### 2025-06-27 #8506 @warrickball
Hi @AbbyBoehm, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Given our current backlog, at the moment I'm afraid this will probably take at least 3 months. Thanks for your patience!

In the meantime, you may want to have a pre-emptive look at our review criteria, particularly on [documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation). From the repo, it appears that there are substantial Sphinx-based docs but I couldn't find them rendered anywhere. From the pages workflow I thought they might be at the corresponding [GitHub pages location](https://exo-tic.github.io/HUSTLE-tools/) (which aren't linked from anywhere, by the way) but that URL only has the project `README.md`. You'll need the docs rendered somewhere before an editor picks this up.
