### 2025-09-05 d175005
various new functions, plot cleanups, and inputs
