### 2025-09-11 f2b8554
new plotting functions, and detrend-based CR rejection
