+++ hustle_tools/stage_1/spatial_outlier_rejection.py
@@ -302,7 +302,7 @@ def laplacian_edge_detection(obs, sigma=10, factor=2, n=2, build_fine_structure=
         obs.data_quality[k] = obs.data_quality[k].where(obs.data_quality[k].values == dq,dq)
 
         if (show_plots == 1 or save_plots == 1) and k == 0:
-            plot_exposure([cumulative_S[0,:,:]], min = 1e-3, max = 1, 
+            plot_exposure([cumulative_S[0,:,:]],# min = 1e-3, max = 1, 
                           show_plot=(show_plots>=1), save_plot=(save_plots>=1), 
                           output_dir=output_dir, filename = ['LED_location_of_corrected_pixels_0'])
             
@@ -311,7 +311,7 @@ def laplacian_edge_detection(obs, sigma=10, factor=2, n=2, build_fine_structure=
                           output_dir=output_dir, filename = ['LED_after_correction_0'])
         
         elif show_plots == 2 or save_plots == 2:
-            plot_exposure([cumulative_S[k,:,:]], min = 1e-3, max = 1, 
+            plot_exposure([cumulative_S[k,:,:]],# min = 1e-3, max = 1, 
                           show_plot=(show_plots==2), save_plot=(save_plots==2), 
                           output_dir=output_dir, filename = ['LED_location_of_corrected_pixels_{}'.format(k)])
             
@@ -321,12 +321,12 @@ def laplacian_edge_detection(obs, sigma=10, factor=2, n=2, build_fine_structure=
             
             if k == 0:
                 # Additionally plot the noise model and fine structure model, if applicable.
-                plot_exposure([noise_model], min = 1e-3, max = 1, 
+                plot_exposure([noise_model,], min = 1e-3, max = 1000, 
                               show_plot=(show_plots==2), save_plot=(save_plots==2), 
                               output_dir=output_dir, filename = ['LED_Noise_Model'])
                 
                 if build_fine_structure:
-                    plot_exposure([F], min = 1e-3, max = 1, 
+                    plot_exposure([F,], min = 1e-3, max = 1000, 
                                   show_plot=(show_plots==2), save_plot=(save_plots==2), 
                                   output_dir=output_dir, filename = ['LED_Fine_Structure_Model'])
     
