### 2025-09-12 5724703
detrend routine moved to stage 2 under TARDIS name
