+++ hustle_tools/read_and_write_config/writer.py
@@ -156,11 +156,6 @@ def Stage1_info():
                                  "replacement",],
                        "Step 3b":["do_free_iter",
                                   "free_sigma"],
-                       "Step 3c":["do_detrend_rej",
-                                  "dtr_sigma",
-                                  "flux_threshold",
-                                  "dtr_window",
-                                  "dtr_replace",],
                        "Step 4a":["do_led",
                                   "led_threshold",
                                   "led_factor",
@@ -216,11 +211,6 @@ def Stage1_info():
                                       "# int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.",],
                            "Step 3b":["# Bool. Whether to use free iteration rejection to clean the timeseries.",
                                       "# float. The sigma to reject outliers at in each iteration. Iterates over each pixel's timeseries until no outliers at this sigma level are found.",],
-                           "Step 3c":["# Bool. Whether to use systematic-detrending rejection to clean the timeseries. Useful for low-cadence datasets where the above two methods are prone to missing trace-striking rays.",
-                                      "# float. The sigma to reject outliers at.",
-                                      "# float. The e-/s counts value above which this method will be used. Method fails if there is insufficient flux to detect trends with, so keep this at least as high as the median wing flux value.",
-                                      "# int. Median time-series trends will be measured using pixels +/- window columns away from target pixel.",
-                                      "# str. If 'tseries', uses median time-series trends to replace outliers. If 'median', uses pixel's median value in time to replace outliers.",],
                            "Step 4a":["# Bool. Whether to use Laplacian Edge Detection rejection to clean the frames.",
                                       "# Float. The threshold parameter at which to kick outliers in LED. The lower the number, the more values will be replaced.",
                                       "# Int. The subsampling factor. Minimum value 2. Higher values increase computation time but aren't expected to yield much improvement in rejection.",
@@ -275,10 +265,11 @@ def Stage2_info():
     subsection_headers = ["# Setup for Stage 2",
                           "# Step 1: Read in the data",
                           "# Step 2: Trace configuration",
-                          "# Step 3: 1D spectral extraction",
-                          "# Step 3a: Box extraction parameters",
-                          "# Step 3b: Optimum extraction parameters",
-                          "# Step 4: 1D spectral cleaning and aligning",
+                          "# Step 3: Time And Relative Detrending In Space (TARDIS)",
+                          "# Step 4: 1D spectral extraction",
+                          "# Step 4a: Box extraction parameters",
+                          "# Step 4b: Optimum extraction parameters",
+                          "# Step 5: 1D spectral cleaning and aligning",
                           ]
     
     subsection_keys = {"Setup":["toplevel_dir",
@@ -291,17 +282,22 @@ def Stage2_info():
                        "Step 2":["path_to_cal",
                                  "traces_to_conf",
                                  "refine_fit"],
-                       "Step 3":["method",
+                       "Step 3":["do_tardis",
+                                 "tardis_sigma",
+                                 "flux_threshold",
+                                 "tardis_window",
+                                 "tardis_replace",],
+                       "Step 4":["method",
                                  "correct_zero",
                                  "subtract_contam",
                                  "sens_correction",
                                  "mask_objs",],
-                       "Step 3a":["determine_hw",
+                       "Step 4a":["determine_hw",
                                   "indices",
                                   "halfwidths_box",],
-                       "Step 3b":["aperture_type",
+                       "Step 4b":["aperture_type",
                                   "halfwidths_opt",],
-                       "Step 4":["outlier_sigma",
+                       "Step 5":["outlier_sigma",
[…]
