+++ hustle_tools/stage_1/temporal_outlier_rejection.py
@@ -231,127 +231,3 @@ def free_iteration_rejection(obs, threshold = 3.5,
 
     return obs
 
-
-def detrend_outlier_rejection(obs, sigma=3.5, flux_threshold=1000, 
-                              window=10, replacement='tseries',
-                              verbose = 0, show_plots = 0, save_plots = 0, output_dir = None):
-    """Estimates systematic trends to catch cosmic rays otherwise not detected
-    by either of the other available time outlier rejection methods.
-
-    Args:
-        obs (xarray): obs.images DataSet contains the images.
-        sigma (float, optional): sigma to reject outliers at. Defaults to 3.5.
-        flux_threshold (float, optional): median e-/s value at which a pixel's
-        time series will be subjected to this method. Method fails if there is
-        insufficient flux by which to estimate systematics. Defaults to 1000.
-        window (int, optional): median time-series of pixels is computed using
-        pixels +/- window length away from target pixel. Defaults to 10.
-        replacement (str, optional): if 'tseries', replaces outliers with scaled
-        median time-series of nearby pixels. If 'median', replace outlier pixels
-        with pixel's median value in time. Defaults to 'tseries'.
-        verbose (int, optional): how detailed you want the printed statements
-        to be. Defaults to 0.
-        show_plots (int, optional): how many plots you want to show. Defaults to 0.
-        save_plots (int, optional): how many plots you want to save. Defaults to 0.
-        output_dir (str, optional): where to save the plots to, if save_plots
-        is greater than 0. Defaults to None.
-
-    Returns:
-        xarray: obs with .images cleaned of CRs and with .data_quality updated
-        to indicate where CRs were found.
-    """
-    # Copy images, get median-normalized t-series, and define hit map.
-    images = obs.images.data.copy()
-    med_images = obs.images.data.copy()
-    med_images /= np.median(med_images,axis=0)
-    hit_map = np.zeros_like(images)
-
-    # Iterate over each suitably-bright pixel.
-    xfix, yfix = np.where(np.median(images,axis=0)>flux_threshold)
-    for x, y in tqdm(zip(xfix,yfix),total=len(xfix),
-                     desc='Identifying CRs in areas with strong systematics... Progress:',
-                     disable=(verbose < 1)):
-        # Do not attempt corrections if y is too close to the detector edge.
-        proximity_alert = (y-window<0) or (y+window>med_images.shape[2])
-        if proximity_alert:
-            continue
-
-        # Get local median time-series.
-        med_series = np.median(med_images[:,x,y-window:y+window],axis=1)
-
-        # Detrend target time-series and measure statistics.
-        detrended_series = med_images[:,x,y]/med_series
-        med = np.median(detrended_series)
-        sig = np.std(detrended_series)
-
-        # Update hit map to record where outliers were found.
-        hit_map[:,x,y] = np.where(np.abs(detrended_series-med)>sigma*sig,1,hit_map[:,x,y])
-    
-    # Report results.
-    if verbose >= 1:
-        print("Identification complete. Pixels identified as %.2f-sigma CRs: %.0f out of %.0f" % (sigma,np.count_nonzero(hit_map),
-                                                                                                  hit_map.shape[0]*hit_map.shape[1]*hit_map.shape[2]))
-
-    # Iterate through hits in time and space and replace as requested.
-    thits, xhits, yhits = np.where(hit_map!=0)
-    for t,x,y in tqdm(zip(thits,xhits,yhits),total=len(thits),
-                      desc='Correcting CRs using {}... Progress:'.format(replacement),
-                      disable=(verbose < 1)):
-        if replacement == 'tseries':
-            # Recompute median time series.
[…]
