+++ hustle_tools/stage_1/__init__.py
@@ -5,7 +5,6 @@ __all__ = [
     "laplacian_edge_detection",
     "spatial_smoothing",
     "fixed_iteration_rejection",
-    "detrend_outlier_rejection",
     "Pagul_bckg_subtraction",
     "column_by_column_subtraction",
     "uniform_value_bkg_subtraction",
@@ -23,6 +22,6 @@ from hustle_tools.stage_1.uvis_embed import get_subarr_coords
 from hustle_tools.stage_1.correct_hst_flags import correct_hst_flags
 from hustle_tools.stage_1.spatial_outlier_rejection import laplacian_edge_detection, spatial_smoothing
 from hustle_tools.stage_1.bckg_subtract import Pagul_bckg_subtraction, uniform_value_bkg_subtraction, column_by_column_subtraction
-from hustle_tools.stage_1.temporal_outlier_rejection import fixed_iteration_rejection, free_iteration_rejection, detrend_outlier_rejection
+from hustle_tools.stage_1.temporal_outlier_rejection import fixed_iteration_rejection, free_iteration_rejection
 from hustle_tools.stage_1.compute_displacements import track_bkgstars, track_0thOrder, refine_location
 from hustle_tools.plotting.plot_exposures import plot_exposure

[…]
