### 2025-09-22 cbf122a
tardis updates and bckg diagnostic plots
