+++ hustle_tools/plotting/plot_bkgsubtraction.py
@@ -216,3 +216,53 @@ def plot_histogram(bin_cents, array, mode, median, exp_num,
     plt.close() # save memory
 
     return
+
+def plot_bkgcorrection(exp_times, pre_bkg, post_bkg, method,
+                       output_dir = None, save_plot = False, show_plot = False):
+    """Plots the pre- and post-subtraction median background flux for the
+    upper right corner of the image.
+
+    Args:
+        exp_times (np.array): BJD exposure times for each frame.
+        pre_bkg (np.array): 1D array of measured background values.
+        post_bkg (np.array): 1D array of corrected background values.
+        method (str): The method used for background subtraction, useful to
+        distinguish each plot file from each other.
+        output_dir (str, optional): output directory where the plot will be
+        saved. Defaults to None.
+        save_plot (bool, optional): whether to save the plot to a file.
+        Defaults to False.
+        show_plot (bool, optional): whether to interrupt execution to
+        show the user the plot. Defaults to False.
+    """
+
+    # initialize figure
+    plt.figure(figsize = (10, 7))
+    # plot values and also median+/-sigma
+    plt.plot(exp_times, pre_bkg, '-o', color='indianred')
+    med, sig = np.median(pre_bkg), np.std(pre_bkg)
+    plt.axhline(med,ls='--',color='indianred')
+    for mult in (-1,1):
+        plt.axhline(med+(mult*sig),ls=':',color='indianred')
+    plt.plot(exp_times, post_bkg, '-o', color='k')
+    med, sig = np.median(post_bkg), np.std(post_bkg)
+    plt.axhline(med,ls='--',color='k')
+    for mult in (-1,1):
+        plt.axhline(med+(mult*sig),ls=':',color='k')
+    plt.xlabel('Exposure')
+    plt.ylabel('Background Counts')
+    plt.title('Raw vs corrected background')
+    
+    if save_plot:
+        plot_dir = os.path.join(output_dir, 'plots') 
+        filedir = os.path.join(plot_dir, 'bkg_correction_{}.png'.format(method))
+        if not os.path.exists(plot_dir):
+            os.makedirs(plot_dir) 
+        plt.savefig(filedir, bbox_inches='tight', dpi=300)
+
+    if show_plot:
+        plt.show(block=True)
+    
+    plt.close() # save memory
+
+    return 
