+++ hustle_tools/plotting/__init__.py
@@ -6,6 +6,7 @@ __all__ = [
     "plot_0th_order",
     "plot_bkgvals",
     "plot_mode_v_params",
+    "plot_bkgcorrection",
     "plot_flags_per_time",
     "plot_one_spectrum",
     "plot_many_spectra",
@@ -27,7 +28,7 @@ __all__ = [
 from hustle_tools.plotting.plot_quicklook import quicklookup
 from hustle_tools.plotting.plot_exposures import plot_exposure
 from hustle_tools.plotting.plot_displacements import plot_bkg_stars, plot_0th_order
-from hustle_tools.plotting.plot_bkgsubtraction import plot_corners, plot_bkgvals, plot_mode_v_params, plot_histogram
+from hustle_tools.plotting.plot_bkgsubtraction import plot_corners, plot_bkgvals, plot_mode_v_params, plot_histogram, plot_bkgcorrection
 from hustle_tools.plotting.plot_timeseries import plot_flags_per_time,  plot_raw_whitelightcurve, plot_raw_spectrallightcurves, plot_aperture_lightcurves, plot_raw_binned_spectrallightcurves, plot_waterfall
 from hustle_tools.plotting.plot_spectra import plot_one_spectrum, plot_many_spectra, plot_spec_gif, plot_2d_spectra
 from hustle_tools.plotting.plot_traces import plot_fitted_amplitudes, plot_fitted_positions, plot_fitted_widths, plot_profile_fit
