+++ hustle_tools/stage_1/bckg_subtract.py
@@ -7,7 +7,7 @@ from scipy.optimize import least_squares
 from scipy.optimize import curve_fit
 from scipy.signal import medfilt, medfilt2d
 
-from hustle_tools.plotting import plot_exposure, plot_corners, plot_bkgvals, plot_mode_v_params, plot_histogram
+from hustle_tools.plotting import plot_exposure, plot_corners, plot_bkgvals, plot_mode_v_params, plot_histogram, plot_bkgcorrection
 
 
 def Pagul_bckg_subtraction(obs, pagul_path, masking_parameter=0.001,
@@ -49,6 +49,14 @@ def Pagul_bckg_subtraction(obs, pagul_path, masking_parameter=0.001,
     # copy images
     images = obs.images.data.copy() 
 
+    # collect pre-subtraction corner flux for reference
+    xshape, yshape = images[0,:,:].shape
+    bx1,bx2 = xshape-25,xshape-1
+    by1,by2 = yshape-500,yshape-1
+    bckg_flux_pre = np.empty(images.shape[0])
+    for i in range(images.shape[0]):
+        bckg_flux_pre[i] = np.median(images[i,bx1:bx2,by1:by2])
+
     # track scaling parameters, should be ~equal to the frame mode
     scaling_parameters = []
     modes, meds = [], []
@@ -61,30 +69,46 @@ def Pagul_bckg_subtraction(obs, pagul_path, masking_parameter=0.001,
 
     # get the subarr_coords
     x1,x2,y1,y2 = [int(x) for x in obs.subarr_coords.values]
-    # iterate over all images
+
+    # build the frame mask using the median frame
+    med_image = np.median(images,axis=0)
+    finite = med_image[np.isfinite(med_image)]
+    bin_number = int(0.10*finite.shape[0])
+    hist, bin_edges = np.histogram(finite, bins=bin_number)
+    ind = np.argmax(hist)
+    mode = (bin_edges[ind]+bin_edges[ind+1])/2
+    sig = np.nanstd(finite)
+    masked_frame = np.ma.masked_where(med_image - mode > masking_parameter*sig, med_image)
+    pagul_mask = masked_frame.mask
+    if verbose > 0:
+        n_masked = 100*np.sum(masked_frame.mask)/(masked_frame.size)
+        print("Percentange of each frame that is masked by this parameter: {:.1f}%".format(n_masked))
+
+    # iterate over all images, applying the same mask
     for k, image in enumerate(tqdm(images, desc = 'Fitting Pagul et al. sky image... Progress:',
                                    disable=(verbose<1))):
         # first, get the coarse frame mode and standard deviation using the frame's finite values
         finite = image[np.isfinite(image)]
-        small = finite[np.abs(finite)<100]
         # pick a reasonable bin_number
-        bin_number = int(0.10*small.shape[0])
-        hist, bin_edges = np.histogram(small, bins=bin_number)
+        bin_number = int(0.10*finite.shape[0])
+        hist, bin_edges = np.histogram(finite, bins=bin_number)
         ind = np.argmax(hist)
         mode = (bin_edges[ind]+bin_edges[ind+1])/2
-        sig = np.nanstd(finite)
+        #sig = np.nanstd(finite)
 
         modes.append(mode)
-        meds.append(np.median(small))
+        meds.append(np.median(finite))
 
-        # next, mask any sources in the frame using the frame mode and standard deviation
-        masked_frame = np.ma.masked_where(np.abs(image - mode) > masking_parameter*sig, image)
+        # next, mask any sources in the frame using the pagul mask
+        #masked_frame = np.ma.masked_where(image - mode > masking_parameter*sig, image)
[…]

[…]
