+++ docs/requirements.txt
@@ -1,0 +1,16 @@
+numpy
+scipy>=1.8.0
+xarray
+astroquery
+astropy
+photutils
+matplotlib
+tqdm
+grismconf
+
+# trying to eliminate these dependencies
+#wfc3tools
+#jwst
+
+# for sphinx documentation
+sphinx_rtd_theme
