### 2025-11-24 6cf8a55
Added explicit path to conf.py
