+++ .readthedocs.yml
@@ -13,3 +13,7 @@ build:
 python:
   install:
     - requirements: docs/requirements.txt
+
+sphinx:
+  # Path to your Sphinx configuration file.
+  configuration: docs/source/conf.py
