### 2025-11-24 1d6b431
Added ipykernel and nbsphinx reqs
