+++ docs/requirements.txt
@@ -9,4 +9,6 @@ tqdm
 grismconf
 wfc3tools
 # for sphinx documentation
+ipykernel
+nbsphinx
 sphinx_rtd_theme
