### 2025-11-24 3030fd0
Update stage_0.rst for hustle_tools rename
