### 2025-11-24 14c2b46
temporary patch to fix readthedocs build
take jwst dependency out once s3-5 are merged
