+++ .readthedocs.yml
@@ -8,7 +8,7 @@ version: 2
 build:
   os: ubuntu-24.04
   tools:
-    python: "3.8"
+    python: "3.11"
 
 python:
   install:
