### 2025-11-24 2863127
Update citation.rst for hustle_tools rename
