+++ docs/source/views/citation.rst
@@ -1,13 +1,13 @@
 Citation
 ========
 
-If you make use of ExoTiC-UVIS in your research, please cite TBD.
+If you make use of HUSTLE-tools in your research, please cite TBD.
 
 .. code-block::
 
         @software{boehm_gascon_23_07_2024,
           author       = {Abby Boehm and Carlos Gascon},
-          title        = {Exo-TiC/ExoTiC-UVIS: ExoTiC-UVIS v0.0.1},
+          title        = {Exo-TiC/HUSTLE-tools: HUSTLE-tools v0.0.1},
           month        = TBD,
           year         = TBD,
           publisher    = {TBD},
