### 2025-11-24 2da6cf6
Update api.rst for hustle_tools rename
