### 2025-11-24 13958ea
Update installation.rst
