### 2025-11-25 f37e223
Update stage_0.rst
trying to get the functions' summaries to show up
