### 2025-11-25 a5532e0
Update stage_0.rst
still trying to get it to show up
