### 2025-11-25 4080e7c
Update stage_0.rst
still trying to get them to show up
