+++ docs/source/views/api/stage_0.rst
@@ -1,17 +1,36 @@
 Stage 0
+=======
+Collect and Move Files
 -------
 
-.. currentmodule:: hustle_tools.stage_0
+.. currentmodule:: hustle_tools.stage_0.collect_and_move_files
 
-.. automodule:: hustle_tools.stage_0
+.. automodule:: hustle_tools.stage_0.collect_and_move_files
 
 .. autosummary::
    :toctree: summary
 
-   hustle_tools.stage_0.collect_and_move_files
+   collect_and_move_files.collect_and_move_files
 
-   hustle_tools.stage_0.identify_orbits
+   collect_and_move_files.identify_orbits
 
-   hustle_tools.stage_0.collect_files
+   collect_and_move_files.collect_files
 
-    
+-------    
+Get Files From Mast
+-------
+
+.. currentmodule:: hustle_tools.stage_0.get_files_from_mast
+
+.. automodule:: hustle_tools.stage_0.get_files_from_mast
+
+.. autosummary::
+   :toctree: summary
+
+   get_files_from_mast.get_files_from_mast
+
+   get_files_from_mast.query_MAST
+
+   get_files_from_mast.download_from_MAST
+
+-------
