### 2025-11-25 f0cdbf5
Update api.rst
tryna get the docs to show up pleaseeeee
