### 2025-11-25 0778f65
Update stage_0.rst
reformatting toc
