### 2025-11-25 ec6caa2
Update api.rst
added other stages
