+++ docs/source/views/api/api.rst
@@ -19,6 +19,28 @@ Stage 0
 Words here.
 
 .. toctree::
-    :maxdepth: 2
+    :maxdepth: 3
 
     stage_0
+
+
+Stage 1
+-------
+
+Words here.
+
+.. toctree::
+    :maxdepth: 3
+
+    stage_1
+
+
+Stage 2
+-------
+
+Words here.
+
+.. toctree::
+    :maxdepth: 3
+
+    stage_2
