### 2025-11-25 c6375f9
Create stage_1.rst
adding all the functions
