+++ docs/source/views/api/stage_1.rst
@@ -1,0 +1,30 @@
+.. currentmodule:: hustle_tools.stage_1
+
+.. automodule:: hustle_tools.stage_1
+
+.. autosummary::
+   :toctree: summary
+
+   hustle_tools.stage_1.load_and_save_data.load_data_S1
+
+   hustle_tools.stage_1.load_and_save_data.save_data_S1
+
+   hustle_tools.stage_1.temporal_outlier_rejection.fixed_iteration_rejection
+
+   hustle_tools.stage_1.temporal_outlier_rejection.free_iteration_rejection
+
+   hustle_tools.stage_1.spatial_outlier_rejection.spatial_smoothing
+
+   hustle_tools.stage_1.spatial_outlier_rejection.laplacian_edge_detection
+
+   hustle_tools.stage_1.bckg_subtract.Pagul_bckg_subtraction
+
+   hustle_tools.stage_1.bckg_subtract.uniform_value_bkg_subtraction
+
+   hustle_tools.stage_1.bckg_subtract.column_by_column_subtraction
+
+   hustle_tools.stage_1.compute_displacements.refine_location
+
+   hustle_tools.stage_1.compute_displacements.track_bkg_stars
+
+   hustle_tools.stage_1.compute_displacements.track_0thOrder
