+++ docs/source/views/api/stage_0.rst
@@ -1,3 +1,6 @@
+Stage 0
+-----------
+
 .. currentmodule:: hustle_tools.stage_0
 
 .. automodule:: hustle_tools.stage_0
