### 2025-11-25 1aa535f
Update stage_0.rst
expanded lists of functions in modules
