### 2025-11-25 aad45d9
Update stage_0.rst
shorter titles, headers
