### 2025-11-25 4f3cd88
Update stage_0.rst
added headers
