### 2025-11-25 bda524a
Update stage_1.rst
headers and sections
