+++ docs/source/index.rst
@@ -1,9 +1,19 @@
-Add a logo
+.. HUSTLE-tools documentation master file.
+
+.. image:: hustle-tools_logo.png
+  :width: 256
+  :align: center
+  :alt: HUSTLE-tools Logo
+
+|
+
+HUSTLE-Tools
+===================================
 
 Introduction
 ------------
 
-ExoTiC-UVIS is a python package for...
+:code:`HUSTLE-tools` is a Python package for downloading, reducing, and extracting 1D spectra from WFC3/UVIS G280 observations. :code:`HUSTLE-tools` is a one stop shop for extracting precise 1D spectral time series from the raw *flt.fits files of your observation using user-friendly and easily-replicable reduction instructions supplied from .hustle configuration files. :code:`HUSTLE-tools` is under active development - stay tuned for future updates which will include tools for exoplanet transmission and emission spectral extraction!
 
 .. toctree::
    :maxdepth: 2
@@ -19,11 +29,11 @@ ExoTiC-UVIS is a python package for...
 Acknowledgements
 ----------------
 
-ExoTiC-UVIS is built by Abby Boehm and Carlos Gascon.
+:code:`HUSTLE-tools` is built and maintained by Abby Boehm and Carlos Gascon.
 
-Add HUSTLE/other contributor acknowledgements here...
+We also thank Hannah Wakeford, David Grant, and Ailsa Campbell for contributing to the development and testing of :code:`HUSTLE-tools`!
 
-If you make use of ExoTiC-UVIS in your research, see the :doc:`citation page <views/citation>`
+If you make use of :code:`HUSTLE-tools` in your research, see the :doc:`citation page <views/citation>`
 for info on how to cite this package. Anyone can get involved by following
 the :doc:`contributing guidelines <views/contributing>`. You can find other
 software from the Exoplanet Timeseries Characterisation (ExoTiC) ecosystem
