### 2025-11-25 4c7e1a0
streamlined api
