### 2025-11-25 838729a
Update api.rst
