+++ docs/source/views/api/api.rst
@@ -12,34 +12,96 @@ hustle_tools.run_pipeline
     :show-inheritance:
 
 
-Stage 0
+Stage 0: Data Handling
 '''''''''''''''''
 
-hustle_tools.stage_0
+Get Files From MAST
 ----------------------------
-.. automodule:: hustle_tools.stage_0
+.. automodule:: hustle_tools.stage_0.get_files_from_mast
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Collect And Move Files
+----------------------------
+.. automodule:: hustle_tools.stage_0.collect_and_move_files
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Locate Target
+----------------------------
+.. automodule:: hustle_tools.stage_0.locate_target
     :members:
     :undoc-members:
     :show-inheritance:
 
 
-Stage 1
+Stage 1: Reduction
 '''''''''''''''''
 
-hustle_tools.stage_1
+Temporal Outlier Rejection
 ----------------------------
-.. automodule:: hustle_tools.stage_1
+.. automodule:: hustle_tools.stage_1.temporal_outlier_rejection
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Spatial Outlier Rejection
+----------------------------
+.. automodule:: hustle_tools.stage_1.spatial_outlier_rejection
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Background Subtraction
+----------------------------
+.. automodule:: hustle_tools.stage_1.bckg_subtract
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Displacement Estimation
+----------------------------
+.. automodule:: hustle_tools.stage_1.compute_displacements
     :members:
     :undoc-members:
     :show-inheritance:
 
 
-Stage 2
+Stage 2: Extraction
 '''''''''''''''''
 
-hustle_tools.stage_2
+Trace Fitting
 ----------------------------
-.. automodule:: hustle_tools.stage_2
+.. automodule:: hustle_tools.stage_2.trace_fitting
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+Extraction
+----------------------------
+.. automodule:: hustle_tools.stage_2.standard_extraction
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. automodule:: hustle_tools.stage_2.optimal_extraction
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+1D Spectral Alignment
+----------------------------
+.. automodule:: hustle_tools.stage_2.align_spectra
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+1D Spectral Cleaning
+----------------------------
+.. automodule:: hustle_tools.stage_2.clean_spectra
     :members:
     :undoc-members:
     :show-inheritance:
