+++ docs/source/views/tutorials.rst
@@ -4,4 +4,6 @@ Tutorials
 .. toctree::
    :maxdepth: 1
 
-   Example tutorial here <tutorials/example_tutorial>
+   Stage 0: Data Handling <tutorials/stage_0>
+   Stage 1: Reduction <tutorials/stage_1>
+   Stage 2: Extraction <tutorials/stage_2>
