### 2025-11-26 e1be7de
Update quick_start.rst
added s0 config
