+++ docs/source/views/quick_start.rst
@@ -11,14 +11,14 @@ The first step to running :code:`HUSTLE-tools` is to make sure you have it and i
 2. Set up a run directory
 -------------------------
 
-We recommend keeping your G280 data reduction projects separate for ease of navigation. For this quickstart, let's create a directory for the observations of HAT-P-41B from visit 01 of HST-GO 15288 (PI: David Sing).:
+We recommend keeping your G280 data reduction projects separate for ease of navigation. For this quickstart, let's create a directory for the observations of HAT-P-41B from visit 01 of HST-GO 15288 (PI: David Sing):
 
 .. code-block:: bash
 
   mkdir /User/hustle-tools_demo/
   cd /User/hustle-tools_demo/
 
-:code:`HUSTLE-tools` operates by reading in .hustle configuration files to a high-level wrapper function, :code:`hustle_tools.run_pipeline()`. To run :code:`HUSTLE-tools` in our new run directory, we'll need to create two items:
+The :code:`HUSTLE-tools` pipeline is operated by reading in .hustle configuration files to a high-level wrapper function, :code:`hustle_tools.run_pipeline()`. To run :code:`HUSTLE-tools` in our new run directory, we'll need to create two items:
   1. A folder to store our .hustle configuration files in.
   2. A simple .py script to run the pipeline with.
 
@@ -54,7 +54,34 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
 
 .. code-block:: bash
 
-  """i will come back to this later i promise :3"""
+  # HUSTLE-tools config file for launching Stage 0: Data Handling
+
+  # Setup for Stage 0
+  toplevel_dir    'output'                                   # Directory where you want your files to be stored after Stage 0 has run. This is where /specimages, /directimages, /visitfiles, and /miscfiles will be stored.
+  verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
+  show_plots      0                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
+  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  
+  # Step 1: Download files from MAST
+  do_download     True                                        # Bool. Whether to perform this step.
+  programID       '15288'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
+target_name     'HAT-P-41B'                                  # Name of the target object you want to query data from. On MAST, referred to as "target_name".
+  token           None                                        # str or None. If you are downloading proprietary data, please visit https://auth.mast.stsci.edu/token?suggested_name=Astroquery&suggested_scope=mast:exclusive_access to obtain an authentication token and enter it as a '' string here.
+  extensions      ['_flt.fits','_spt.fits']                   # lst of str or None. File extensions you want to download. If None, take all file extensions. Otherwise, take only the files specified. _flt.fits, _spt.fits recommended as minimum working case.
+  
+  # Step 2: Organizing files
+  do_organize     True                                        # Bool. Whether to perform this step.
+  visit_number    '01'                                        # The visit number you want to operate on.
+  filesfrom_dir   None                                        # None or str. If you downloaded data in Step 1, leave this as None. If you have pre-downloaded data, please place all of it in filesfrom_dir. Don't sort it into sub-folders; HUSTLE-tools won't be able to find them if they are inside sub-folders!
+  
+  # Step 3: Locating the target star
+  do_locate       True                                        # Bool. Whether to perform this step.
[…]
