### 2025-11-26 4458af8
Update quick_start.rst
added s1 and s2 configs
