+++ docs/source/views/quick_start.rst
@@ -48,7 +48,7 @@ That's it!
 2.2. Supply the configuration files
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The .hustle configuration files are at the core of operating :code:`HUSTLE-tools` and they take some time to get to know. For this quickstart, we've written the .hustle files for you, but if you want to learn more about how to tune these files for your needs, check out the :ref:`Tutorials` tab! For now, just :code:`cd` into :code:`configs` and follow the instructions below to create the .hustle configuration files for this run.
+The .hustle configuration files are at the core of operating :code:`HUSTLE-tools` and they take some time to get to know. For this quickstart, we've written most of the .hustle files for you, but if you want to learn more about how to tune these files for your needs, check out the :ref:`Tutorials` tab! For now, just :code:`cd` into :code:`configs` and follow the instructions below to create the .hustle configuration files for this run.
 
 First, use your favorite text editor to create :code:`configs/stage_0_input_config.hustle` and populate it with the following script:
 
@@ -65,7 +65,7 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
   # Step 1: Download files from MAST
   do_download     True                                        # Bool. Whether to perform this step.
   programID       '15288'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
-target_name     'HAT-P-41B'                                  # Name of the target object you want to query data from. On MAST, referred to as "target_name".
+  target_name     'HAT-P-41B'                                  # Name of the target object you want to query data from. On MAST, referred to as "target_name".
   token           None                                        # str or None. If you are downloading proprietary data, please visit https://auth.mast.stsci.edu/token?suggested_name=Astroquery&suggested_scope=mast:exclusive_access to obtain an authentication token and enter it as a '' string here.
   extensions      ['_flt.fits','_spt.fits']                   # lst of str or None. File extensions you want to download. If None, take all file extensions. Otherwise, take only the files specified. _flt.fits, _spt.fits recommended as minimum working case.
   
@@ -87,13 +87,131 @@ Next, create :code:`configs/stage_1_input_config.hustle` and populate it with th
 
 .. code-block:: bash
 
-  """this too :3"""
+  # HUSTLE-tools config file for launching Stage 1: Reduction
+  
+  # Setup for Stage 1
+  toplevel_dir    'output'                                   # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
+  output_run      'demo'                                     # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
+  show_plots      0                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
+  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  
+  # Step 1: Read in the data
+  skip_first_fm   True                                        # Bool. If True, ignores all first frames in each orbit.
+  skip_first_or   False                                       # Bool. If True, ignores all frames in the first orbit.
+  
+  # Step 2: Reject cosmic rays with time iteration
+  # Step 2a: Fixed iteration parameters
+  do_fixed_iter   True                                        # Bool. Whether to use fixed iteration rejection to clean the timeseries.
+  fixed_sigmas    [3.5,3.5]                                   # lst of float. The sigma to reject outliers at in each iteration. The length of the list is the number of iterations.
+  replacement     7                                           # int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.
+  
[…]
