### 2025-11-26 09a20c1
Update quick_start.rst
most of the quickstart is in place now
