+++ docs/source/views/quick_start.rst
@@ -1,12 +1,12 @@
 Quick start
 ===========
 
-Ready to get going with :code:`HUSTLE-tools`? This quickstart guide will help you understand the basics of running :code:`HUSTLE-tools` on your G280 data. More detailed instructions for each Stage can be found in the :ref:`Tutorials` page.
+Ready to get going with :code:`HUSTLE-tools`? This quickstart guide will help you understand the basics of running :code:`HUSTLE-tools` on your G280 data. More detailed instructions for each Stage can be found in the `Tutorials <https://hustle-tools.readthedocs.io/en/latest/views/tutorials.html>`_ page.
 
 1. Install :code:`HUSTLE-tools`
 -------------------------------
 
-The first step to running :code:`HUSTLE-tools` is to make sure you have it and its dependencies installed. Follow the instructions on the :ref:`Installation` page to get your :code:`HUSTLE-tools` conda environment set up and ready to go.
+The first step to running :code:`HUSTLE-tools` is to make sure you have it and its dependencies installed. Follow the instructions on the `Installation <https://hustle-tools.readthedocs.io/en/latest/views/installation.html>`_ page to get your :code:`HUSTLE-tools` conda environment set up and ready to go.
 
 2. Set up a run directory
 -------------------------
@@ -18,16 +18,7 @@ We recommend keeping your G280 data reduction projects separate for ease of navi
   mkdir /User/hustle-tools_demo/
   cd /User/hustle-tools_demo/
 
-The :code:`HUSTLE-tools` pipeline is operated by reading in .hustle configuration files to a high-level wrapper function, :code:`hustle_tools.run_pipeline()`. To run :code:`HUSTLE-tools` in our new run directory, we'll need to create two items:
-  1. A folder to store our .hustle configuration files in.
-  2. A simple .py script to run the pipeline with.
-
-.. code-block:: bash
-
-  mkdir configs
-  vim run_pipeline.py
-
-We'll need to populate each of these items as follows.
+The :code:`HUSTLE-tools` pipeline is operated by reading in .hustle configuration files to a high-level wrapper function, :code:`hustle_tools.run_pipeline()`. To run :code:`HUSTLE-tools` in our new run directory, we'll need to create a folder for the configuration files, and a script to run :code:`hustle_tools.run_pipeline()` with. Let's call our configuration file folder :code:`/User/hustle-tools_demo/configs`, and our pipeline script :code:`/User/hustle-tools_demo/run_pipeline.py`.
 
 2.1. Create the pipeline wrapper script
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -48,7 +39,7 @@ That's it!
 2.2. Supply the configuration files
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-The .hustle configuration files are at the core of operating :code:`HUSTLE-tools` and they take some time to get to know. For this quickstart, we've written most of the .hustle files for you, but if you want to learn more about how to tune these files for your needs, check out the :ref:`Tutorials` tab! For now, just :code:`cd` into :code:`configs` and follow the instructions below to create the .hustle configuration files for this run.
+The .hustle configuration files are at the core of operating :code:`HUSTLE-tools` and they take some time to get to know. For this quickstart, we've written most of the .hustle files for you, but if you want to learn more about how to tune these files for your research, check out the `Tutorials <https://hustle-tools.readthedocs.io/en/latest/views/tutorials.html>`_ tab! For now, just :code:`cd` into :code:`configs` and follow the instructions below to create the .hustle configuration files for this run.
 
 First, use your favorite text editor to create :code:`configs/stage_0_input_config.hustle` and populate it with the following script:
 
[…]
