+++ docs/source/index.rst
@@ -10,7 +10,7 @@
 HUSTLE-Tools: A One-Stop Shop for WFC3/UVIS G280 Data Reduction
 ===============================================================
 
-:code:`HUSTLE-tools` is a one stop shop Python Package for extracting precise 1D spectral time series from WFC3/UVIS G280 time-series obserations. :code:`HUSTLE-tools` processes the raw *flt.fits files of your observation with a user-friendly interface and easily-replicable reduction instructions supplied from .hustle configuration files. :code:`HUSTLE-tools` is under active development - stay tuned for future updates which will include tools for exoplanet transmission and emission spectral extraction!
+:code:`HUSTLE-tools` is a one-stop shop Python package for extracting precise 1D spectral time series from WFC3/UVIS G280 time-series obserations. :code:`HUSTLE-tools` processes the raw \*flt.fits files of your observation with a user-friendly interface and easily-replicable reduction instructions supplied from .hustle configuration files. :code:`HUSTLE-tools` is under active development - stay tuned for future updates which will include tools for exoplanet transmission and emission spectrum fitting!
 
 .. toctree::
    :maxdepth: 2
@@ -30,8 +30,4 @@ Acknowledgements
 
 We thank Hannah Wakeford, David Grant, and Ailsa Campbell for contributing to the development and testing of :code:`HUSTLE-tools`!
 
-If you make use of :code:`HUSTLE-tools` in your research, see the :doc:`citation page <views/citation>`
-for info on how to cite this package. Anyone can get involved by following
-the :doc:`contributing guidelines <views/contributing>`. You can find other
-software from the Exoplanet Timeseries Characterisation (ExoTiC) ecosystem
-over on `GitHub <https://github.com/Exo-TiC>`_.
+If you make use of :code:`HUSTLE-tools` in your research, see the :doc:`citation page <views/citation>` for info on how to cite this package. Anyone can get involved by following the :doc:`contributing guidelines <views/contributing>`. You can find other software from the Exoplanet Timeseries Characterisation (ExoTiC) ecosystem over on `GitHub <https://github.com/Exo-TiC>`_.
