### 2025-11-26 3a2aeaf
Update installation.rst
