+++ docs/source/views/installation.rst
@@ -26,4 +26,4 @@ _______________________________________________________
 Step 3 (optional): Download WFC3/UVIS sky background files
 __________________________________________________________
 
-The WFC3/UVIS G280 has a unique background pattern owing in part to the amplifier split midway across the detector (see `Pagul et al. 2023 <https://www.stsci.edu/files/live/sites/www/files/home/hst/instrumentation/wfc3/documentation/instrument-science-reports-isrs/_documents/2023/WFC3-ISR-2023-06.pdf>`_ for details). :code:`HUSTLE-tools` uses the empirically-determined G280 sky image in a custom background correction routine contained in Step 5c of Stage 1. If you want to make use of this routine, simply head over to `https://www.stsci.edu/hst/instrumentation/wfc3/documentation/grism-resources/uvis-grism-sky-images <https://www.stsci.edu/hst/instrumentation/wfc3/documentation/grism-resources/uvis-grism-sky-images>`_ and download the latest type FLT sky image appropriate to the chip used in your observations (typically chip 2). Then supply the absolute path to this file on your computer in the Stage 1 .hustle file.
+The WFC3/UVIS G280 has a unique background pattern owing in part to the amplifier split midway across the detector (see `Pagul et al. 2023 <https://www.stsci.edu/files/live/sites/www/files/home/hst/instrumentation/wfc3/documentation/instrument-science-reports-isrs/_documents/2023/WFC3-ISR-2023-06.pdf>`_ for details). :code:`HUSTLE-tools` uses the empirically-determined G280 sky image developed by `Pagul et al. 2023 <https://www.stsci.edu/files/live/sites/www/files/home/hst/instrumentation/wfc3/documentation/instrument-science-reports-isrs/_documents/2023/WFC3-ISR-2023-06.pdf>`_ in a custom background correction routine contained in Step 5c of Stage 1. If you want to make use of this routine, simply head over to `https://www.stsci.edu/hst/instrumentation/wfc3/documentation/grism-resources/uvis-grism-sky-images <https://www.stsci.edu/hst/instrumentation/wfc3/documentation/grism-resources/uvis-grism-sky-images>`_ and download the latest sky image appropriate to the file type and chip used in your observations (typically type FLT, chip 2). Then supply the absolute path to this file on your computer in the Stage 1 .hustle file.
