### 2025-11-26 0e69dbc
Update quick_start.rst
