+++ docs/source/views/quick_start.rst
@@ -48,7 +48,7 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
   # HUSTLE-tools config file for launching Stage 0: Data Handling
 
   # Setup for Stage 0
-  toplevel_dir    'output'                                   # Directory where you want your files to be stored after Stage 0 has run. This is where /specimages, /directimages, /visitfiles, and /miscfiles will be stored.
+  toplevel_dir    'output'                                    # Directory where you want your files to be stored after Stage 0 has run. This is where /specimages, /directimages, /visitfiles, and /miscfiles will be stored.
   verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
   show_plots      0                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
   save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
@@ -56,7 +56,7 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
   # Step 1: Download files from MAST
   do_download     True                                        # Bool. Whether to perform this step.
   programID       '15288'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
-  target_name     'HAT-P-41B'                                  # Name of the target object you want to query data from. On MAST, referred to as "target_name".
+  target_name     'HAT-P-41B'                                 # Name of the target object you want to query data from. On MAST, referred to as "target_name".
   token           None                                        # str or None. If you are downloading proprietary data, please visit https://auth.mast.stsci.edu/token?suggested_name=Astroquery&suggested_scope=mast:exclusive_access to obtain an authentication token and enter it as a '' string here.
   extensions      ['_flt.fits','_spt.fits']                   # lst of str or None. File extensions you want to download. If None, take all file extensions. Otherwise, take only the files specified. _flt.fits, _spt.fits recommended as minimum working case.
   
@@ -81,8 +81,8 @@ Next, create :code:`configs/stage_1_input_config.hustle` and populate it with th
   # HUSTLE-tools config file for launching Stage 1: Reduction
   
   # Setup for Stage 1
-  toplevel_dir    'output'                                   # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
-  output_run      'demo'                                     # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  toplevel_dir    'output'                                    # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
+  output_run      'demo'                                      # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
   verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
   show_plots      0                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
   save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
@@ -233,7 +233,7 @@ The outputs of the pipeline will be stored in :code:`output/outputs/`. Stages 0,
 4.1. The outputs of Stage 0
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
[…]
