### 2025-11-26 0a4bd93
Update and rename example_tutorial.rst to stage_0.rst
stage 0 tutorial setup
