+++ docs/source/views/tutorials/example_tutorial.rst
@@ -1,4 +1,0 @@
-Example tutorial
-================
-
-TBD.
