+++ docs/source/views/tutorials/stage_0.rst
@@ -1,0 +1,90 @@
+Stage 0: Data Handling
+======================
+
+The first stage of any data reduction project is to download the data. :code:`HUSTLE-tools` Stage 0 handles downloading and organizing the data for your program ID, visit number, and target of interest. The only engagement required from you is to supply the .hustle configuration file and identify the target star in the direct image. This tutorial will walk you through this process.
+
+Creating the Stage 0 configuration file
+---------------------------------------
+
+The first step is to create the configuration file that will guide the execution of Stage 0. Create a folder to store the configuration file in, e.g. :code:`configs/`. Then create :code:`configs/stage_0_input_config.hustle` and populate it with the following:
+
+.. code-block:: bash
+
+  # HUSTLE-tools config file for launching Stage 0: Data Handling
+
+  # Setup for Stage 0
+  toplevel_dir    'output'                                    # Directory where you want your files to be stored after Stage 0 has run. This is where /specimages, /directimages, /visitfiles, and /miscfiles will be stored.
+  verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
+  show_plots      2                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
+  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  
+  # Step 1: Download files from MAST
+  do_download     True                                        # Bool. Whether to perform this step.
+  programID       '12345'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
+  target_name     'PLANET-NAME'                               # Name of the target object you want to query data from. On MAST, referred to as "target_name".
+  token           None                                        # str or None. If you are downloading proprietary data, please visit https://auth.mast.stsci.edu/token?suggested_name=Astroquery&suggested_scope=mast:exclusive_access to obtain an authentication token and enter it as a '' string here.
+  extensions      ['_flt.fits','_spt.fits']                   # lst of str or None. File extensions you want to download. If None, take all file extensions. Otherwise, take only the files specified. _flt.fits, _spt.fits recommended as minimum working case.
+  
+  # Step 2: Organizing files
+  do_organize     True                                        # Bool. Whether to perform this step.
+  visit_number    '01'                                        # The visit number you want to operate on.
+  filesfrom_dir   None                                        # None or str. If you downloaded data in Step 1, leave this as None. If you have pre-downloaded data, please place all of it in filesfrom_dir. Don't sort it into sub-folders; HUSTLE-tools won't be able to find them if they are inside sub-folders!
+  
+  # Step 3: Locating the target star
+  do_locate       True                                        # Bool. Whether to perform this step.
+  location        None                                        # None or tuple of float. Prior to running Stage 0, this will be None. After running Stage 0, a copy of this .hustle file will be made with this information included.
+  
+  # Step 4: Quality quicklook
+  do_quicklook    True                                        # Bool. Whether to perform this step.
+  
+  # ENDPARSE
+
+Let's break down each of these steps to make sure we understand what we can tune in this stage.
+
+Setup for Stage 0
+~~~~~~~~~~~~~~~~~
[…]
