### 2025-11-26 260dc2e
Create stage_1.rst
starting the tutorial
