+++ docs/source/views/tutorials/stage_1.rst
@@ -1,0 +1,146 @@
+Stage 1: Reduction
+======================
+
+**NOTE: Make sure you ran Stage 0: Data Handling before attempting this stage!**
+  
+With the data in hand, it is time to clean the data of cosmic rays, hot pixels, and background signal. :code:`HUSTLE-tools` Stage 1 takes the files downloaded in Stage 1 and attempts to correct sources of noise to improve the signal-to-noise ratio of your extracted spectra. This tutorial will walk you through this process.
+
+Creating the Stage 1 configuration file
+---------------------------------------
+
+The first step is to create the configuration file that will guide the execution of Stage 1. Create a folder to store the configuration file in, e.g. :code:`configs/`. Then create :code:`configs/stage_1_input_config.hustle` and populate it with the following:
+
+.. code-block:: bash
+
+  # HUSTLE-tools config file for launching Stage 1: Reduction
+  
+  # Setup for Stage 1
+  toplevel_dir    'output'                                    # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
+  output_run      'run1'                                      # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
+  show_plots      2                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
+  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  
+  # Step 1: Read in the data
+  skip_first_fm   False                                       # Bool. If True, ignores all first frames in each orbit.
+  skip_first_or   False                                       # Bool. If True, ignores all frames in the first orbit.
+  
+  # Step 2: Reject cosmic rays with time iteration
+  # Step 2a: Fixed iteration parameters
+  do_fixed_iter   True                                        # Bool. Whether to use fixed iteration rejection to clean the timeseries.
+  fixed_sigmas    [10,10]                                     # lst of float. The sigma to reject outliers at in each iteration. The length of the list is the number of iterations.
+  replacement     7                                           # int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.
+  
+  # Step 2b: Free iteration parameters
+  do_free_iter    False                                       # Bool. Whether to use free iteration rejection to clean the timeseries.
+  free_sigma      3.5                                         # float. The sigma to reject outliers at in each iteration. Iterates over each pixel's timeseries until no outliers at this sigma level are found.
+  
+  # Step 3: Reject hot pixels with spatial detection
+  # Step 3a: Laplacian Edge Detection parameters
+  do_led          True                                        # Bool. Whether to use Laplacian Edge Detection rejection to clean the frames.
+  led_threshold   5                                           # Float. The threshold parameter at which to kick outliers in LED. The lower the number, the more values will be replaced.
+  led_factor      2                                           # Int. The subsampling factor. Minimum value 2. Higher values increase computation time but aren't expected to yield much improvement in rejection.
+  led_n           2                                           # Int. Number of times to do LED on each frame. Enter None to continue performing LED on each frame until no outliers are found.
+  fine_structure  True                                        # Bool. Whether to build a fine structure model, which can protect narrow bright features like traces from LED.
+  contrast_factor 5                                           # Float. If fine_structure is True, acts as the led_threshold for the fine structure step.
+  
[…]
