### 2025-11-26 8ade025
Update stage_1.rst
