### 2025-11-26 9732c71
Update stage_1.rst
