+++ docs/source/views/tutorials/stage_1.rst
@@ -116,15 +116,34 @@ Fixed iteration rejection iterates over every pixel's time series a fixed number
 Step 2b: Free iteration parameters
 ''''''''''''''''''''''''''''''''''
 Free iteration rejection iterates over every pixel's time series at a single rejection threshold, and will continue iterating at that threshold until no outliers are found. Set :code:`do_free_iter` to True if you want to use this method. The threshold that will be used on these iterations is set as a float supplied to the :code:`free_sigma` variable.
+
+Step 3: Reject hot pixels with spatial detection
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Hot pixels output too much signal compared to their neighbors, while cold/dead pixels output little to no signal compared to their neighbors. Both contribute additiona noise to extracted spectra but neither are substantially time-varying and therefore cannot be treated with time-based methods. :code:`HUSTLE-tools` offers two routines for treating hot, cold, and dead pixels: Laplacian Edge Detection and spatial smoothing.
+
+Step 3a: Laplacian Edge Detection parameters
+''''''''''''''''''''''''''''''''''''''''''''
+Laplacian Edge Detection (LED) is a method of detecting defective pixels through their steep changes in flux levels relative to their neighbors. It was developed and detailed in `van Dokkum 2001 <https://iopscience.iop.org/article/10.1086/323894>`_ and we implement this method into :code:`HUSTLE-tools`. Set :code:`do_led` to True if you want to apply this routine to your G280 data. Adjust :code:`led_threshold` to set the threshold at which outliers are rejected, with lower values flagging more pixels. You can also adjust the :code:`led_factor` to change the factor by which the frames are subsampled during LED (see `van Dokkum 2001 <https://iopscience.iop.org/article/10.1086/323894>`_ for details). You can set the number of iterations LED is run for by adjusting :code:`led_n`, which may be useful for whittling down large blobs of hot, cold, and/or dead pixels since only the edge pixels will be detected on each iteration. Spectral traces are thin structures which can have sharp edges that can be mistakenly attacked by LED. To protect large fine features such as traces from excessive targeting by LED, set :code:`fine_structure` to True and adjust the :code:`contrast_factor` to tune how sensitive this model is to rejecting outliers. When :code:`fine_structure` is True, only pixels that trip both the :code:`led_threshold` and :code:`contrast_factor` thresholds will be rejected.
+
+Step 3b: Spatial smoothing parameters
+'''''''''''''''''''''''''''''''''''''
+LED can be time-intensive as it requires building a noise model and optionally a fine structure model for each full frame interpolated to higher resolution. :code:`HUSTLE-tools` also includes a simple spatial median filtering method which is faster because it operates on native resolution frames and can operate on limited frame regions, but may attack fine features more aggressively. Set :code:`do_smooth` to True to apply this routine to your G280 data. select one of three :code:`smth_type` options ('1D_smooth', '2D_smooth', 'polyfit') to adjust how the median filtered model of each frame is built. The window or kernel size used for filtering is set by :code:`smth_kernel` and ideally should be similar or less than the lengthscale of your trace width. Adjust :code:`smth_threshold` to higher/lower values to flag outliers less/more aggressively. To speed reduction by focusing only on the data (particularly useful if the 'polyfit' method is used), set :code:`smth_bounds` to encompass only the traces you intend to extract from.
+
+Step 4: Background subtraction
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
[…]
