### 2025-11-26 4f9c9e1
Update stage_1.rst
