### 2025-12-05 2e28a63
Update stage_1.rst
