+++ docs/source/views/tutorials/stage_1.rst
@@ -111,7 +111,7 @@ Cosmic rays, or energetic particles striking the detector at random and originat
 
 Step 2a: Fixed iteration parameters
 '''''''''''''''''''''''''''''''''''
-Fixed iteration rejection iterates over every pixel's time series a fixed number of times, using a specified threshold at each iteration to reject outliers. Set :code:`do_fixed_iter` to True if you want to use this method. The number of iterations, and the threshold for rejection at each iteration, is specified by supplying a list of thresholds to the :code:`fixed_sigmas` variable. When outliers are found, their handling is controlled by the :code:`replacement` variable, which can be set to an integer to specify the half-width of the running median window used to compute the replacement value, or can be set to None to replace the outlier by the overall median of the pixel's time series.
+Fixed iteration rejection iterates over every pixel's time series a fixed number of times, using a specified threshold at each iteration to reject outliers. Set :code:`do_fixed_iter` to True if you want to use this method. The number of iterations, and the threshold for rejection at each iteration, is specified by supplying a list of thresholds to the :code:`fixed_sigmas` variable. When outliers are found, their handling is controlled by the :code:`replacement` variable, which can be set to an integer to specify the half-width of the running median window used to compute the replacement value, or can be set to None to replace the outlier by the median of the pixel's full time series.
 
 Step 2b: Free iteration parameters
 ''''''''''''''''''''''''''''''''''
@@ -119,7 +119,7 @@ Free iteration rejection iterates over every pixel's time series at a single rej
 
 Step 3: Reject hot pixels with spatial detection
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Hot pixels output too much signal compared to their neighbors, while cold/dead pixels output little to no signal compared to their neighbors. Both contribute additiona noise to extracted spectra but neither are substantially time-varying and therefore cannot be treated with time-based methods. :code:`HUSTLE-tools` offers two routines for treating hot, cold, and dead pixels: Laplacian Edge Detection and spatial smoothing.
+Hot pixels output too much signal compared to their neighbors, while cold/dead pixels output little to no signal compared to their neighbors. Both contribute additional noise to extracted spectra but neither are substantially time-varying and therefore cannot be treated with time-based methods. :code:`HUSTLE-tools` offers two routines for treating hot, cold, and dead pixels: Laplacian Edge Detection and spatial smoothing.
 
 Step 3a: Laplacian Edge Detection parameters
 ''''''''''''''''''''''''''''''''''''''''''''
@@ -131,7 +131,7 @@ LED can be time-intensive as it requires building a noise model and optionally a
 
 Step 4: Background subtraction
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-The HST WFC3/UVIS G280 sky is complicated, showing variations in both space and time. The trace also occupies a substantial fraction of the frame, and the slitles nature of the G280 grism means that background objects' dispersed spectra can interfere with background estimation. To accomodate this, :code:`HUSTLE-tools` offers a wide range of background treatment options for you to explore to determine which suits your data best.
+The HST WFC3/UVIS G280 sky is complicated, showing variations in both space and time. The trace also occupies a substantial fraction of the frame, and the slitless nature of the G280 grism means that background objects' dispersed spectra can interfere with background estimation. To accomodate this, :code:`HUSTLE-tools` offers a wide range of background treatment options for you to explore to determine which suits your data best.
 
 Step 4a: uniform value background subtraction
 '''''''''''''''''''''''''''''''''''''''''''''
@@ -143,12 +143,11 @@ This method of background correction treats the value of the background as const
   
 Step 4c: Pagul et al. background subtraction
[…]
