### 2025-12-05 cb11d5e
Create stage_2.rst
