+++ docs/source/views/tutorials/stage_2.rst
@@ -1,0 +1,112 @@
+Stage 2: Extraction
+======================
+
+**NOTE: Make sure you ran Stage 0: Data Handling and Stage 1: Reduction before attempting this stage!**
+  
+Now that the data are clean, we want to extract the 1D spectral time series from the traces. :code:`HUSTLE-tools` Stage 2 calibrates the wavelength solution of your cleaned data to extract the spectrum from 200 to 800 nm. This tutorial will walk you through this process.
+
+Creating the Stage 2 configuration file
+---------------------------------------
+
+The first step is to create the configuration file that will guide the execution of Stage 2. Create a folder to store the configuration file in, e.g. :code:`configs/`. Then create :code:`configs/stage_2_input_config.hustle` and populate it with the following:
+
+.. code-block:: bash
+
+  # HUSTLE-tools config file for launching Stage 2: Extraction
+  
+  # Setup for Stage 2
+  toplevel_dir    'output'                                    # Directory where your current project files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data as well as the outputs folder.
+  input_run       'run_1'                                     # Str. This is the name of the Stage 1 run you want to load.
+  output_run      'run_1'                                     # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
+  show_plots      2                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
+  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  
+  # Step 1: Read in the data
+  
+  # Step 2: Trace configuration
+  path_to_cal     './'                                        # Str. The absolute path to the .conf file used by GRISMCONF for the chip your data were taken on.
+  traces_to_conf  ('+1','-1')                                 # Lst of str. The traces you want to configure and extraction from.
+  refine_fit      False                                       # Bool. If True, uses Gaussian fitting to refine the trace solution.
+  
+  # Step 3: 1D spectral extraction
+  method          'box'                                       # Str. Options are 'box' (draw a box around the trace and sum without weights) or 'optimal' (weight using Horne 1986 methods).
+  correct_zero    False                                       # Bool. Whether to model the contaminating 0th order and subtract it from your data during extraction. Sometimes works, sometimes just adds lots of scatter.
+  subtract_contam False                                       # Bool. Whether to model the contaminating orders and subtract them from your trace during extraction. Sometimes works, sometimes just adds lots of scatter.
+  sens_correction False                                       # Bool. Whether to correct for the G280's changing sensitivity as a function of wavelength. Since absolute calibrated spectra aren't needed in exoplanetary sciences, you can skip this safely.
+  mask_objs       []                                          # List of lists. If there are background objects in your planned aperture, mask them here. Each entry is (x,y,radius).
+  
+  # Step 3a: Box extraction parameters
+  determine_hw    False                                       # Bool. If True, automatically determines preferred half-width for each order by minimizing out-of-transit/eclipse residuals.
+  indices         ([0,10],[-10,-1])                           # Lst of lsts of int. If determine_hw, these are the indices used to estimate the out-of-transit/eclipse residuals.
+  halfwidths_box  (12,12)                                     # Lst of ints. The half-width of extraction aperture to use for each order. Input here is ignored if 'determine_hw' is True.
[…]
