### 2025-12-05 389210e
Update stage_2.rst
added some pictures
