### 2025-12-05 d53a96a
Update stage_2.rst
fixed spacing
