### 2025-12-08 72a7ce8
Update setup.py
fixed photutils version req
