+++ docs/source/views/quick_start.rst
@@ -11,7 +11,7 @@ The first step to running :code:`HUSTLE-tools` is to make sure you have it and i
 2. Set up a run directory
 -------------------------
 
-We recommend keeping your G280 data reduction projects separate for ease of navigation. For this quick start, let's create a directory for the observations of HAT-P-41B from visit 01 of HST-GO 15288 (PI: David Sing):
+We recommend keeping your G280 data reduction projects separate for ease of navigation. For this quick start, let's create a directory for the HUSTLE program observations of WASP-127b from visit 12 of HST-GO 17183 (PI: Hannah Wakeford):
 
 .. code-block:: bash
 
@@ -55,14 +55,14 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
   
   # Step 1: Download files from MAST
   do_download     True                                        # Bool. Whether to perform this step.
-  programID       '15288'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
-  target_name     'HAT-P-41B'                                 # Name of the target object you want to query data from. On MAST, referred to as "target_name".
+  programID       '17183'                                     # ID of the observing program you want to query data from. On MAST, referred to as "proposal_ID".
+  target_name     'WASP-127'                                  # Name of the target object you want to query data from. On MAST, referred to as "target_name".
   token           None                                        # str or None. If you are downloading proprietary data, please visit https://auth.mast.stsci.edu/token?suggested_name=Astroquery&suggested_scope=mast:exclusive_access to obtain an authentication token and enter it as a '' string here.
   extensions      ['_flt.fits','_spt.fits']                   # lst of str or None. File extensions you want to download. If None, take all file extensions. Otherwise, take only the files specified. _flt.fits, _spt.fits recommended as minimum working case.
   
   # Step 2: Organizing files
   do_organize     True                                        # Bool. Whether to perform this step.
-  visit_number    '01'                                        # The visit number you want to operate on.
+  visit_number    '12'                                        # The visit number you want to operate on.
   filesfrom_dir   None                                        # None or str. If you downloaded data in Step 1, leave this as None. If you have pre-downloaded data, please place all of it in filesfrom_dir. Don't sort it into sub-folders; HUSTLE-tools won't be able to find them if they are inside sub-folders!
   
   # Step 3: Locating the target star
@@ -94,7 +94,7 @@ Next, create :code:`configs/stage_1_input_config.hustle` and populate it with th
   # Step 2: Reject cosmic rays with time iteration
   # Step 2a: Fixed iteration parameters
   do_fixed_iter   True                                        # Bool. Whether to use fixed iteration rejection to clean the timeseries.
-  fixed_sigmas    [3.5,3.5]                                   # lst of float. The sigma to reject outliers at in each iteration. The length of the list is the number of iterations.
+  fixed_sigmas    [5.0,5.0]                                   # lst of float. The sigma to reject outliers at in each iteration. The length of the list is the number of iterations.
   replacement     7                                           # int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.
   
   # Step 2b: Free iteration parameters
@@ -121,7 +121,7 @@ Next, create :code:`configs/stage_1_input_config.hustle` and populate it with th
   # Step 4a: uniform value background subtraction
   do_uniform      True                                        # Bool. Whether to subtract the background using one uniform value as the value for the entire frame.
[…]
