### 2025-12-08 1892b09
Update quick_start.rst
