### 2025-12-08 5edd122
Update installation.rst
