### 2025-12-08 597d47e
Update stage_0.rst
