### 2025-12-08 0352b15
Create stage_0_output.txt
