+++ docs/source/views/tutorials/stage_0_output.txt
@@ -1,0 +1,791 @@
+Querying MAST for files under program ID 17183, target WASP-127...
+Found 169 files related to program ID 17183, target WASP-127.
+Creating directory output to store your files in...
+Downloaded data will be stored in directory output.
+Examining 251 queried data products for files in visit number 12...
+Downloading file number 0 from mast:HST/product/iexr12fhq_spt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fhq_spt.fits to output/iexr12fhq_spt.fits ...
+|================================================================================================================================| 103k/103k (100.00%)         0s
+COMPLETE : None
+Downloading file number 1 from mast:HST/product/iexr12fhq_flt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fhq_flt.fits to output/iexr12fhq_flt.fits ...
+|================================================================================================================================|  12M/ 12M (100.00%)         0s
+COMPLETE : None
+Downloading file number 2 from mast:HST/product/iexr12fdq_spt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fdq_spt.fits to output/iexr12fdq_spt.fits ...
+|================================================================================================================================| 103k/103k (100.00%)         0s
+COMPLETE : None
+Downloading file number 3 from mast:HST/product/iexr12fdq_flt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fdq_flt.fits to output/iexr12fdq_flt.fits ...
+|================================================================================================================================|  12M/ 12M (100.00%)         1s
+COMPLETE : None
+Downloading file number 4 from mast:HST/product/iexr12fgq_spt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fgq_spt.fits to output/iexr12fgq_spt.fits ...
+|================================================================================================================================| 103k/103k (100.00%)         0s
+COMPLETE : None
+Downloading file number 5 from mast:HST/product/iexr12fgq_flt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fgq_flt.fits to output/iexr12fgq_flt.fits ...
+|================================================================================================================================|  12M/ 12M (100.00%)         0s
+COMPLETE : None
+Downloading file number 6 from mast:HST/product/iexr12fqq_spt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fqq_spt.fits to output/iexr12fqq_spt.fits ...
+|================================================================================================================================| 103k/103k (100.00%)         0s
+COMPLETE : None
+Downloading file number 7 from mast:HST/product/iexr12fqq_flt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12fqq_flt.fits to output/iexr12fqq_flt.fits ...
+|================================================================================================================================|  12M/ 12M (100.00%)         1s
+COMPLETE : None
+Downloading file number 8 from mast:HST/product/iexr12f7q_spt.fits...
+Downloading URL https://mast.stsci.edu/api/v0.1/Download/file?uri=mast:HST/product/iexr12f7q_spt.fits to output/iexr12f7q_spt.fits ...
+|================================================================================================================================| 103k/103k (100.00%)         0s
+COMPLETE : None
[…]
