### 2025-12-08 9c3fc27
Update stage_0.rst
