### 2025-12-08 aa6e9e4
Update stage_1.rst
