### 2025-12-08 b7e6603
Update stage_2.rst
