### 2025-12-09 95d3627
Update stage_1.rst
