### 2025-12-09 1b469fe
Update stage_0.rst
