### 2025-12-09 f91d6eb
Update stage_2.rst
