### 2026-01-12 #9723 @AbbyBoehm
@radicamc @ivastar Hello all! The docs do exist on a separate branch, I'm waiting to merge them into main pending a second reviewer to approve the merge. They are mostly complete, pending the Citation page and some big touch-ups to the API section. You can find them hosted here: https://hustle-tools.readthedocs.io/en/latest/

I am not sure what has caused them to not show up on Google, but once I've figured it out and fixed it I will let you know!
