### 2026-02-09 43b638d
fixed input to be wavelengths, not pixels
