### 2026-02-09 1655058
note to make wavelength range variable
