### 2026-02-09 86e970f
saturation blooming
