### 2026-02-09 8cf4b57
align and apply_align vars
