### 2026-02-09 d04ecb8
align and remove zero updates
