+++ hustle_tools/run_pipeline.py
@@ -337,7 +337,7 @@ def run_pipeline(config_files_dir, stages=(0, 1, 2, 3, 4, 5)):
             # test 0th order removal
             remove_zeroth_order(obs, 
                                 zero_pos = [x0th, y0th], 
-                                rmin = 100, rmax = 300, rwidth = 3, 
+                                rmin = 85, rmax = 500, rwidth = 3,
                                 fit_profile = True,
                                 verbose = stage2_dict['verbose'],
                                 show_plots = stage2_dict['show_plots'],
@@ -423,8 +423,9 @@ def run_pipeline(config_files_dir, stages=(0, 1, 2, 3, 4, 5)):
                                                         spec,
                                                         spec_err,
                                                         order,
-                                                        trace_x=wav,
-                                                        align=True,
+                                                        trace_x=trace_x,#wav,
+                                                        wavelengths=wav,
+                                                        align=stage2_dict['apply_align'],
                                                         verbose=stage2_dict['verbose'],
                                                         show_plots=stage2_dict['show_plots'], 
                                                         save_plots=stage2_dict['save_plots'],
