### 2026-02-09 2ab65e3
quickstart notebook created, all congruent
