+++ docs/source/views/quick_start.rst
@@ -66,8 +66,8 @@ First, use your favorite text editor to create :code:`configs/stage_0_input_conf
   filesfrom_dir   None                                        # None or str. If you downloaded data in Step 1, leave this as None. If you have pre-downloaded data, please place all of it in filesfrom_dir. Don't sort it into sub-folders; HUSTLE-tools won't be able to find them if they are inside sub-folders!
   
   # Step 3: Locating the target star
-  do_locate       True                                        # Bool. Whether to perform this step.
-  location        None                                        # None or tuple of float. Prior to running Stage 0, this will be None. After running Stage 0, a copy of this .hustle file will be made with this information included.
+  do_locate       False                                       # Bool. Whether to perform this step.
+  location        [974.8154860553503, 160.15959552212985]     # None or tuple of float. Prior to running Stage 0, this will be None. After running Stage 0, a copy of this .hustle file will be made with this information included.
   
   # Step 4: Quality quicklook
   do_quicklook    True                                        # Bool. Whether to perform this step.
@@ -85,7 +85,7 @@ Next, create :code:`configs/stage_1_input_config.hustle` and populate it with th
   output_run      'demo'                                      # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
   verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
   show_plots      0                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
-  save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
+  save_plots      1                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
   
   # Step 1: Read in the data
   skip_first_fm   True                                        # Bool. If True, ignores all first frames in each orbit.
@@ -207,14 +207,14 @@ Lastly, create :code:`configs/stage_2_input_config.hustle` and populate it with 
 3. Run :code:`HUSTLE-tools`
 --------------------------
 
-You are now ready to run :code:`HUSTLE-tools`!
+You are now ready to run :code:`HUSTLE-tools`! Run the following command line script to start processing your dataset:
 
 .. code-block::
 
   cd ~/User/hustle-tools_demo/
   python run_pipeline.py
 
-Most of the pipeline will run hands-free. However, in Stage 0 you will be presented with the direct photometric image taken as part of these observations and asked to locate the target star in the image, which is essential to getting the wavelength solution right. In these observations, you will find the target star at :code:`x=974.82` and :code:`y=160.16`. After this step, the pipeline will operate on its own. On an average laptop with a good internet connection, it should take about ten minutes.
+Once started, the pipeline will operate hands-free. On an average laptop with good wi-fi (~100 Mbps), it should take about 10 minutes for this notebook to process.
 
 4. Examine the outputs
 ----------------------
@@ -228,7 +228,7 @@ Now let's check out the products of each stage. All of our outputs will have bee
   3. :code:`output/visitfiles/` contains files that were associated with the program ID, visit, and specific orbit, but not associated with image data. For this dataset, no visit files were identified.
   4. :code:`output/miscfiles/` contains all other files associated with the program ID and visit.
 
[…]
