+++ tests/test_runpipeline.py
@@ -1,4 +1,0 @@
-from hustle_tools import run_pipeline
-
-
-run_pipeline()
