### 2026-02-09 fa960cb
updated API for plotting
