### 2026-02-09 7980286
api now includes plot and config backend functions
