### 2026-02-09 df5507f
stage 0 api updates
