### 2026-02-09 c626e88
api fixes for stage 1
