+++ hustle_tools/stage_2/load_and_save_data.py
@@ -9,8 +9,7 @@ def load_data_S2(data_dir, filename = 'clean_obs'):
 
     Args:
         data_dir (str): where the reduced .nc file from Stage 1 is kept.
-        filename (str, optional): name of the reduced .nc file.
-        Defaults to 'clean_obs'.
+        filename (str, optional): name of the reduced .nc file. Defaults to 'clean_obs'.
 
     Returns:
         xarray: reduced observations xarray to extract from.
@@ -25,26 +24,20 @@ def save_data_S2(obs, spec, spec_err,
                  trace_x, trace_y, widths, wavelengths,
                  spec_disp, prof_disp, order = '+1',
                  output_dir = None, filename = 'specs'):
-    """Function to create and save xarray containing the information extracted
-    from stage 2.
+    """Function to create and save xarray containing the information extracted from stage 2.
 
     Args:
-        obs (xarray): reduced observations xarray, only needed now for its
-        exp_time data.
+        obs (xarray): reduced observations xarray, only needed now for its exp_time data.
         specs (list): each extracted spectrum.
         specs_err (list): each extracted spectrum's uncertainties.
         trace_x (list): each extracted spectrum's dispersion solution.
         trace_y (list): each extracted spectrum's spatial solutions.
         widths (list): the width of each extracted trace.
         wavelengths (list): each extracted spectrum's wavelength solution.
-        spec_shifts (list): each extracted spectrum's dispersion and
-        spatial shifts.
-        orders (tuple, optional): which orders are being saved.
-        Defaults to ("+1", "-1").
-        output_dir (str, optional): where to save the files to.
-        Defaults to None.
-        filename (str, optional): name to give each file, joined to its order
-        string. Defaults to 'specs'.
+        spec_shifts (list): each extracted spectrum's dispersion and spatial shifts.
+        orders (tuple, optional): which orders are being saved. Defaults to ("+1", "-1").
+        output_dir (str, optional): where to save the files to. Defaults to None.
+        filename (str, optional): name to give each file, joined to its order string. Defaults to 'specs'.
     """
 
     # Create and save xarray for each order
