+++ hustle_tools/stage_2/trace_fitting.py
@@ -14,8 +14,7 @@ from hustle_tools.plotting import plot_fitted_positions
 
 def get_calibration_0th(obs, source_pos, path_to_cal,
                         verbose = 0, show_plots = 0, save_plots = 0, output_dir = None):
-    """Uses a custom method based on GRISMCONF and the source position to
-    locate the 0th order for removal purposes.
+    """Uses a custom method based on GRISMCONF and the source position to locate the 0th order for removal purposes.
 
     Args:
         x0 (float): Embedded x position of the source.
@@ -101,31 +100,21 @@ def get_calibration_0th(obs, source_pos, path_to_cal,
 
 def get_trace_solution(obs, order, source_pos, refine_calibration, path_to_cal,
                        verbose = 0, show_plots = 0, save_plots = 0, output_dir = None):
-    """Pulls the region of each image that has the trace in it, with wavelength
-    solution provided by GRISMCONF.
+    """Pulls the region of each image that has the trace in it, with wavelength solution provided by GRISMCONF.
 
     Args:
-        obs (xarray): obs.images DataSet contains the images and obs.subarr_coords
-        DataSet is used to trick the configuration into thinking it is embedded.
-        order (str): options are "+1", "-1", "+2", "-2", etc. Which order you
-        want to pull.
-        source_pos (tup): x, y float position of the source in the unembedded
-        direct image.
-        refine_calibration (bool): if True, uses Gaussian fitting to improve
-        the location of the trace.
-        path_to_cal (str): path to the GRISMCONF calibration file used to
-        locate the trace.
-        verbose (int, optional): How detailed you want the printed statements
-        to be. Defaults to 0.
+        obs (xarray): obs.images DataSet contains the images and obs.subarr_coords DataSet is used to trick the configuration into thinking it is embedded.
+        order (str): options are "+1", "-1", "+2", "-2", etc. Which order you want to pull.
+        source_pos (tup): x, y float position of the source in the unembedded direct image.
+        refine_calibration (bool): if True, uses Gaussian fitting to improve the location of the trace.
+        path_to_cal (str): path to the GRISMCONF calibration file used to locate the trace.
+        verbose (int, optional): How detailed you want the printed statements to be. Defaults to 0.
         show_plots (int, optional): How many plots you want to show. Defaults to 0.
         save_plots (int, optional): How many plots you want to save. Defaults to 0.
-        output_dir (str, optional): Where to save the plots to, if save_plots
-        is greater than 0. Defaults to None.
+        output_dir (str, optional): Where to save the plots to, if save_plots is greater than 0. Defaults to None.
 
     Returns:
-        np.array,np.array,np.array,np.array,scipy.interpolate: trace x and y
-        positions, wavelength solutions, optional trace widths, and optional
-        sensitivity correction functions.
+        np.array,np.array,np.array,np.array,scipy.interpolate: trace x and y positions, wavelength solutions, optional trace widths, and optional sensitivity correction functions.
     """
    
     # Get the mean subarr_coords offsets.
@@ -183,19 +172,16 @@ def get_trace_solution(obs, order, source_pos, refine_calibration, path_to_cal,
 
 
 def get_calibration_trace(order, x0, y0, path_to_cal):
-    """Uses the supplied calibration software and source position to locate the
-    trace and assign wavelength solution.
+    """Uses the supplied calibration software and source position to locate the trace and assign wavelength solution.
 
     Args:
-        order (str): Options are "+1", "-1", "+2", "-2", etc. Used to grab the
-        right calibration from the calibration file.
+        order (str): Options are "+1", "-1", "+2", "-2", etc. Used to grab the right calibration from the calibration file.
[…]

[…]
