### 2026-02-10 f790ebb
10 million tiny bug fixes, also a few new funcs
