+++ hustle_tools/stage_2/load_and_save_data.py
@@ -1,6 +1,5 @@
 import os
 
-import numpy as np
 import xarray as xr
 
 
@@ -46,6 +45,7 @@ def save_data_S2(obs, spec, spec_err,
             spec = (['exp_time', 'x'], spec),
             spec_err = (['exp_time', 'x'], spec_err),
             trace = (['exp_time', 'x'], trace_y),
+            orbit_numbers = (['exp_time',], obs.orbit_numbers.data),
             ),
         coords=dict(
             wave=(['x'], wavelengths),
@@ -60,9 +60,7 @@ def save_data_S2(obs, spec, spec_err,
         spectra['prof_disp'] = xr.DataArray(prof_disp, dims=['exp_time', 'x']) 
     if widths is not False:
         spectra['fit_widths'] = xr.DataArray(widths, dims=['exp_time', 'x']) 
-        
-    #for i in range(len(bkg_stars)):
-    #    spectra['stars{}_disp'.format(i + 1)] = obs['star{}_disp'.format(i)]   
+    
     spectra['meanstar_disp'] = obs['meanstar_disp']
 
     # Save results in Stage 2 folder 
