+++ hustle_tools/run_pipeline.py
@@ -1,25 +1,25 @@
 import os
 import glob
 import numpy as np
-import matplotlib.pyplot as plt
 
 from hustle_tools.read_and_write_config import parse_config
 from hustle_tools.read_and_write_config import write_config
 
 from hustle_tools.plotting import plot_one_spectrum
+from hustle_tools.plotting import plot_many_spectra
 from hustle_tools.plotting import plot_spec_gif
 from hustle_tools.plotting import plot_2d_spectra
 from hustle_tools.plotting import plot_raw_whitelightcurve
-from hustle_tools.plotting import plot_raw_spectrallightcurves
 from hustle_tools.plotting import quicklookup
 
 from hustle_tools.stage_0 import collect_and_move_files
 from hustle_tools.stage_0 import get_files_from_mast
 from hustle_tools.stage_0 import locate_target
-from hustle_tools.stage_0 import check_spt_subarray
+from hustle_tools.stage_0 import check_subarray
 
 from hustle_tools.stage_1 import load_data_S1
 from hustle_tools.stage_1 import save_data_S1
+from hustle_tools.stage_1 import correct_hst_flags
 from hustle_tools.stage_1 import uniform_value_bkg_subtraction
 from hustle_tools.stage_1 import Pagul_bckg_subtraction
 from hustle_tools.stage_1 import column_by_column_subtraction
@@ -36,6 +36,7 @@ from hustle_tools.stage_2 import save_data_S2
 from hustle_tools.stage_2 import get_calibration_0th
 from hustle_tools.stage_2 import get_trace_solution
 from hustle_tools.stage_2 import sens_correct
+from hustle_tools.stage_2 import time_and_relative_detrending_in_space
 from hustle_tools.stage_2 import determine_ideal_halfwidth
 from hustle_tools.stage_2 import standard_extraction
 from hustle_tools.stage_2 import optimal_extraction
@@ -44,16 +45,13 @@ from hustle_tools.stage_2 import align_spectra
 from hustle_tools.stage_2 import align_profiles
 from hustle_tools.stage_2 import remove_zeroth_order
 
-#from hustle_tools.stage_3 import load_data_S3
-#from hustle_tools.stage_3 import save_data_S3
 
-
-def run_pipeline(config_files_dir, stages=(0, 1, 2, 3, 4, 5)):
+def run_pipeline(config_files_dir, stages=(0, 1, 2)):
     """Wrapper for all Stages of the HUSTLE-tools pipeline.
 
     Args:
         config_files_dir (str): folder which contains the .hustle files needed to run the stages you want to run.
-        stages (tuple, optional): the stages that you want to run. Defaults to (0, 1, 2, 3, 4, 5).
+        stages (tuple, optional): the stages that you want to run. Defaults to (0, 1, 2).
     """
     ######## Run Stage 0 ########
     if 0 in stages:
@@ -86,8 +84,8 @@ def run_pipeline(config_files_dir, stages=(0, 1, 2, 3, 4, 5)):
         # locate target in direct image
         if stage0_dict['do_locate']:
             # check for direct image / spec image discrepancies
-            xdiscs, ydiscs = check_spt_subarray(os.path.join(stage0_dict['toplevel_dir'],'directimages/or01dr001_spt.fits'),
-                                                sorted(glob.glob(os.path.join(os.path.join(stage0_dict['toplevel_dir'],'specimages'),'*spt.fits'))))
+            xdiscs, ydiscs = check_subarray(os.path.join(stage0_dict['toplevel_dir'],'directimages/or01dr001_flt.fits'),
+                                                sorted(glob.glob(os.path.join(os.path.join(stage0_dict['toplevel_dir'],'specimages'),'*flt.fits'))))
[…]
