+++ hustle_tools/plotting/plot_timeseries.py
@@ -1,4 +1,5 @@
 import os
+
 import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib.cm as cm
@@ -131,53 +132,6 @@ def plot_raw_whitelightcurve(times, spec, order="+1",
     return 
 
 
-def plot_raw_spectrallightcurves(times, spec, order="+1",
-                                 show_plot = False, save_plot = False,
-                                 filename = None, output_dir = None):
-    """Plots the uncorrected spectrally-binned light curves for this order, as diagnostics of your cleaning process.
-
-    Args:
-        times (np.array): mid-exposure time of each frame.
-        spec (np.array): 1D extracted spectra.
-        order (str, optional): which order we are plotting, for plot title. Defaults to "+1".
-        show_plot (bool, optional): whether to interrupt execution to show the user the plot. Defaults to False.
-        save_plot (bool, optional): whether to save this plot to a file. Defaults to False.
-        filename (str, optional): name to give this file, if saving. Defaults to None.
-        output_dir (str, optional): where to save the file, if saving. Defaults to None.
-    """
-
-    # define order colors
-    colors = {"+1":'indianred',"-1":'dodgerblue',
-              "+2":'orangered',"-2":'royalblue',
-              "+3":'darkorange',"-3":'blue',
-              "+4":'orange',"-4":'deepskyblue'}    
-
-    for i, lc in enumerate(np.transpose(spec)):
-        n_oot = int(0.20*lc.shape[0]) # typically, first 20% of data is the first orbit, which is oot/ooe
-        raw_lc = lc/np.median(lc[:n_oot])
-
-        plt.figure(figsize = (10, 7))
-        plt.plot(times, raw_lc, 'o', color=colors[order], markeredgecolor='black')
-        plt.xlabel('Time of exposure')
-        plt.ylabel('Counts')
-        plt.title("{}th column's spectral light curve, order {}".format(i,order))
-
-      
-        if save_plot:
-            plot_dir = os.path.join(output_dir, 'plots') 
-            if not os.path.exists(plot_dir):
-                os.makedirs(plot_dir) 
-            filedir = os.path.join(plot_dir, f'{filename}_lc{i}.png')
-            plt.savefig(filedir,dpi=300,bbox_inches='tight')
-
-        if show_plot:
-            plt.show(block=True)
-
-        plt.close() # save memory
-
-    return 
-
-
 def plot_aperture_lightcurves(obs, tested_hws, wlcs,  
                               show_plot = False, save_plot = False,
                               filename = None, output_dir = None):
@@ -196,9 +150,6 @@ def plot_aperture_lightcurves(obs, tested_hws, wlcs,
     # colormap
     cmap = cm.get_cmap('viridis')
     cs = cmap(np.linspace(0,1,len(tested_hws)))
-
-    # offsets
-    #offsets = np.arange(0, len(tested_hws))*0.001
 
     plt.figure(figsize=(10, 7))
     for wlc, hw, c in zip(wlcs, tested_hws, cs):
