+++ hustle_tools/stage_3/__init__.py
@@ -1,9 +1,0 @@
-__all__ = [
-    "bin_light_curves",
-    "load_data_S3",
-    "save_data_S3"
-]
-
-from hustle_tools.stage_3.binning import bin_light_curves
-from hustle_tools.stage_3.load_and_save_data import load_data_S3
-from hustle_tools.stage_3.load_and_save_data import save_data_S3
