### 2026-02-10 aae6f47
updates to the config files and places they are used
