+++ docs/source/views/tutorials/stage_2.rst
@@ -14,41 +14,49 @@ The first step is to create the configuration file that will guide the execution
 .. code-block:: bash
 
   # HUSTLE-tools config file for launching Stage 2: Extraction
-  
+
   # Setup for Stage 2
-  toplevel_dir    'output'                                    # Directory where your current project files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data as well as the outputs folder.
-  input_run       'run1'                                      # Str. This is the name of the Stage 1 run you want to load.
-  output_run      'run1'                                      # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  toplevel_dir    './files'                                   # Directory where your current project files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data as well as the outputs folder.
+  input_run       'run_1'                                     # Str. This is the name of the Stage 1 run you want to load.
+  output_run      'run_1'                                     # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
   verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
   show_plots      2                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
   save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
-  
+
   # Step 1: Read in the data
-  
+
   # Step 2: Trace configuration
   path_to_cal     './'                                        # Str. The absolute path to the .conf file used by GRISMCONF for the chip your data were taken on.
   traces_to_conf  ('+1','-1')                                 # Lst of str. The traces you want to configure and extraction from.
-  refine_fit      False                                       # Bool. If True, uses Gaussian fitting to refine the trace solution.
-  
-  # Step 3: 1D spectral extraction
+  refine_fit      True                                        # Bool. If True, uses Gaussian fitting to refine the trace solution.
+
+  # Step 3: Time And Relative Detrending In Space (TARDIS)
+  do_tardis       True                                        # Bool. Whether to use TARDIS to clean the timeseries. Useful for low-cadence datasets where Stage 1 cleaning methods are prone to missing trace-striking cosmic rays.
+  tardis_sigma    [5.0,3.5]                                   # list of float. The sigmas to reject outliers at. One sigma per window value is needed.
+  flux_threshold  1000                                        # float. The e-/s counts value above which this method will be used. Method fails if there is insufficient flux to detect trends with, so keep this at least as high as the median wing flux value.
+  tardis_window   [10,5]                                      # list of int. Median time-series trends will be measured using pixels +/- window columns away from target pixel.
+  tardis_replace  'tseries'                                   # str. If 'tseries', uses median time-series trends to replace outliers. If 'median', uses pixel's median value in time to replace outliers.
+
+  # Step 4: 1D spectral extraction
   method          'box'                                       # Str. Options are 'box' (draw a box around the trace and sum without weights) or 'optimal' (weight using Horne 1986 methods).
   correct_zero    False                                       # Bool. Whether to model the contaminating 0th order and subtract it from your data during extraction. Sometimes works, sometimes just adds lots of scatter.
[…]
