+++ docs/source/views/tutorials/stage_1.rst
@@ -14,86 +14,94 @@ The first step is to create the configuration file that will guide the execution
 .. code-block:: bash
 
   # HUSTLE-tools config file for launching Stage 1: Reduction
-  
+
   # Setup for Stage 1
-  toplevel_dir    'output'                                    # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
-  output_run      'run1'                                      # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
+  toplevel_dir    './files'                                   # Directory where your Stage 0 files are stored. This folder should contain the specimages/, directimages/, etc. folders with your data.
+  output_run      'run_1'                                     # Str. This is the name to save the current run to. It can be anything that does not contain spaces or special characters (e.g. $, %, @, etc.).
   verbose         2                                           # Int from 0 to 2. 0 = print nothing. 1 = print some statements. 2 = print every action.
   show_plots      2                                           # Int from 0 to 2. 0 = show nothing. 1 = show some plots. 2 = show all plots.
   save_plots      2                                           # Int from 0 to 2. 0 = save nothing. 1 = save some plots. 2 = save all plots.
-  
+
   # Step 1: Read in the data
   skip_first_fm   False                                       # Bool. If True, ignores all first frames in each orbit.
   skip_first_or   False                                       # Bool. If True, ignores all frames in the first orbit.
-  
-  # Step 2: Reject cosmic rays with time iteration
-  # Step 2a: Fixed iteration parameters
+
+  # Step 2: Correct pixels flagged by HST pipeline
+  do_hst_flags    True                                        # Bool. If True, reads in HST data quality information and corrects for chosen HST flags.
+  hst_flags       [16,4096]                                   # List of int. Which HST flags to correct. See HST WFC3/UVIS G280 handbook for flag meanings.
+  hst_replace     True                                        # Bool. If True, replaces HST-flagged pixels with spatial median (if flag<4096) or temporal median (if flag>=4096). Otherwise, adds HST-flagged pixels to badpix_map.
+
+  # Step 3: Reject cosmic rays with time iteration
+  # Step 3a: Fixed iteration parameters
   do_fixed_iter   True                                        # Bool. Whether to use fixed iteration rejection to clean the timeseries.
   fixed_sigmas    [10,10]                                     # lst of float. The sigma to reject outliers at in each iteration. The length of the list is the number of iterations.
-  replacement     7                                           # int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.
-  
-  # Step 2b: Free iteration parameters
-  do_free_iter    False                                       # Bool. Whether to use free iteration rejection to clean the timeseries.
+  replacement     None                                        # int or None. If int, replaces flagged outliers with the median of values within +/-replacement indices of the outlier. If None, uses the median of the whole timeseries instead.
+
+  # Step 3b: Free iteration parameters
+  do_free_iter    True                                        # Bool. Whether to use free iteration rejection to clean the timeseries.
   free_sigma      3.5                                         # float. The sigma to reject outliers at in each iteration. Iterates over each pixel's timeseries until no outliers at this sigma level are found.
-  
-  # Step 3: Reject hot pixels with spatial detection
-  # Step 3a: Laplacian Edge Detection parameters
+
+  # Step 4: Reject hot pixels with spatial detection
+  # Step 4a: Laplacian Edge Detection parameters
   do_led          True                                        # Bool. Whether to use Laplacian Edge Detection rejection to clean the frames.
[…]

[…]
