### 2026-02-10 9a11647
bringing joss material over from joss-md branch
