+++ joss/paper.bib
@@ -216,3 +216,12 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Gascon2025,
+      title={The HUSTLE Program: The UV to Near-IR Transmission Spectrum of the Hot Jupiter KELT-7b}, 
+      author={Carlos Gascón and Mercedes López-Morales and Ryan J. MacDonald and Joanna K. Barstow and Victoria A. Boehm and Hannah R. Wakeford and Munazza K. Alam and Lili Alderson and Natasha E. Batalha and Charlotte E. Fairman and David Grant and Nikole K. Lewis and Mark S. Marley and Sarah E. Moran and Kazumasa Ohno and Guillem Anglada-Escudé and Ignasi Ribas},
+      year={2025},
+      eprint={2506.19193},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.EP},
+      url={https://arxiv.org/abs/2506.19193}, 
+}
