### 2026-02-10 416ded5
tutorials all done, plot touch-ups
