+++ hustle_tools/plotting/__init__.py
@@ -28,4 +28,4 @@ from hustle_tools.plotting.plot_displacements import plot_bkg_stars, plot_0th_or
 from hustle_tools.plotting.plot_bkgsubtraction import plot_corners, plot_bkgvals, plot_mode_v_params, plot_histogram, plot_bkgcorrection
 from hustle_tools.plotting.plot_timeseries import plot_flags_per_time,  plot_raw_whitelightcurve, plot_aperture_lightcurves
 from hustle_tools.plotting.plot_spectra import plot_one_spectrum, plot_many_spectra, plot_spec_gif, plot_2d_spectra
-from hustle_tools.plotting.plot_traces import plot_fitted_amplitudes, plot_fitted_positions, plot_fitted_widths, plot_profile_fit
+from hustle_tools.plotting.plot_traces import plot_fitted_positions, plot_profile_fit
