### 2026-02-23 8c7ef9b
fixed plotting bugs in s2
