### 2026-02-26 #9723 @AbbyBoehm
@ivastar @radicamc @mberg3 Hello all and thanks for your patience! Updates were recently pushed to main which address the following:

1. All documentation for HUSTLE-tools is now hosted from the main branch at: https://hustle-tools.readthedocs.io/en/main/
2. Tutorials between the readthedocs and the GitHub repository have been brought mostly into alignment - there is a small discrepancy in that the tutorials on GitHub are hosted in .ipynb notebook form and can be fully run from within a Jupyter notebook kernel, while the readthedocs tutorials are designed around a terminal-based workflow in which the pipeline is executed via a .py script. Both are valid workflows for using HUSTLE-tools and come down to user preference. They are also linked to each other so that the user can easily find and use the other workflow if they prefer.
3. Likewise, the quickstart guide on readthedocs now has a .ipynb equivalent on GitHub, and the two are mostly congruent except that the readthedocs quickstart is structured around .py/terminal workflows while the GitHub quickstart is for notebook users. The quickstart is very on-rails with a definite deliverable and most of the legwork done already for the user, while the tutorials are open-ended and designed to allow the user to explore and customize the pipeline steps to their preference. The recommended order to follow is as they are presented on the readthedocs page, i.e. beginning with the quickstart and then moving on to the tutorials.
4. A few bug fixes have been pushed through, most importantly a patch on setup.py that restricts the installed photutils version to photutils<=1.13.0 which prevents DAOStarFinder from failing to import properly.
5. Docstrings edits have been made which should ensure the API on readthedocs formats properly.
6. The joss folder including the paper has also been added to the main branch.

In addition, two data reduction functions that were in development have been pushed to main in Stages 1 and 2. The Stage 1 function `correct_hst_flags()` can be used to replace pixels flagged by the HST calwf3 pipeline. The Stage 2 function `time_and_relative_detrending_in_space()` is designed for cleaning trace-striking cosmic rays which are missed by Stage 1 due to strong HST systematics. Both of these functions are included in the Stages 1 and 2 tutorials as well as the API.

Many thanks for the feedback so far! Hopefully this patch addresses all previously-raised concerns well and review can proceed solely on the main branch, but if you have any further concerns please send them my way and we will patch the pipeline as needed.
