### 2026-03-02 #9723 @AbbyBoehm
@ivastar We had substantial assistance from Ailsa Marie Campbell (ORCID [0009-0006-4198-719X]) in the review of the recent pull request and as a result we were wondering if it would be possible to add this person as sixth author on our paper submission? If that is okay with you, we will update the JOSS markdown files on our repository as needed.
