### 2026-04-12 #9723 @mberg3
@AbbyBoehm I have completed my review.

1) For the paper, you focus on the codes available for the WCF3/IR. I would add GRIZLI to that list. Additionally, you did not mention any of the codes for WFC3/UVIS, which include HSTaXe from STScI (highly out of date) and an IDL code (must purchase software license to run) described in J. M. O’Meara et al. (2011) and E. Lusso et al. (2018). Explaining why these codes are not good options for the community would strengthen your statement of need.

2) I think it would be good to add how to access the data and show a plot that the user can use to verify their test run against yours. I am not familiar with xarray or nc files, so I had to find the function necessary to open the data (though not difficult). Since this code is reducing the spectra, having a results plot somewhere would also showcase the science-ready output.

3) Is there a magnitude and/or S/N limit where this code is no longer able to detect the object and its traces? When is an object too bright where you have to start taking the higher orders into account? You mention the Prizkal GRISMCONF files can trace up to order 4. Does your code only focus on and extract the 1st order, or can it also go up to order 4? Does it coadd these orders, or do you output all of them?

4) Did you have any issues with reduction, tracing, and/or extraction of the -1 order, since it has lower S/N than +1?

5) This is a short paper. I am actually very interested in how you dealt with the cosmic rays, shifts, and especially the background correction. Perhaps you are saving this for a journal article.
