### 2026-04-12 #9723 @radicamc
Ha! What timing! @AbbyBoehm I have also completed my review.
Thanks for dealing with my previous comments and streamlining the tutorials and documentation. I have a few additional comments, mostly similar to things that @mberg3 has already raised.
1. In the software paper, you could also mention exoTEDRF-UVIS (Radica et al. 2025)  --- ultimately a repurposed JWST code so your specialized code is more ideal.
2. Like @mberg3 I'm not familiar with either xarray or nc files. Maybe in the QuickStart, you could at least link the relevant documentation, or even better, provide a couple of lines of code showing how to access the major relevant quantities.
3. The tutorials you have included are nice. However, I would prefer to directly see the final output plots/results of your runs. This allows for much easier comparison instead of trying to interpret myself whether "The calibration and aperture plots accurately located the correct traces and the apertures encompass the entirety of the orders you have chosen to extract from." (from Stage 2). For a first time UVIS user, it might not be obvious which is the +1, -1, etc. orders. So explicitly showing things like this would be helpful.
4. Please link the documentation directly in the GitHub ReadMe.
5. Petition to change the name back to hazelnut ;) That was top tier!
