### 2026-04-22 06d02ad
added refs to grizli, idl, and hstaxe
