+++ joss/paper.bib
@@ -1,4 +1,63 @@
+@software{grizli,
+       author = {{Brammer}, Gabriel},
+        title = "{grizli}",
+         year = 2023,
+        month = sep,
+          eid = {10.5281/zenodo.8370018},
+          doi = {10.5281/zenodo.8370018},
+      version = {1.9.11},
+    publisher = {Zenodo},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023zndo...8370018B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
+@ARTICLE{idl1,
+       author = {{O'Meara}, John M. and {Prochaska}, J. Xavier and {Chen}, Hsiao-Wen and {Madau}, Piero},
+        title = "{The Advanced Camera for Surveys+Wide Field Camera 3 Survey for Lyman Limit Systems. I. The Data}",
+      journal = {\apjs},
+     keywords = {intergalactic medium, methods: data analysis, quasars: absorption lines, quasars: general, surveys, ultraviolet: general},
+         year = 2011,
+        month = aug,
+       volume = {195},
+       number = {2},
+          eid = {16},
+        pages = {16},
+          doi = {10.1088/0067-0049/195/2/16},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ApJS..195...16O},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{idl2,
+       author = {{Lusso}, Elisabeta and {Fumagalli}, Michele and {Rafelski}, Marc and {Neeleman}, Marcel and {Prochaska}, Jason X. and {Hennawi}, Joseph F. and {O'Meara}, John M. and {Theuns}, Tom},
+        title = "{The Spectral and Environment Properties of z {\ensuremath{\sim}} 2.0-2.5 Quasar Pairs}",
+      journal = {\apj},
+     keywords = {accretion, accretion disks, galaxies: active, intergalactic medium, quasars: absorption lines, quasars: general, Astrophysics - Astrophysics of Galaxies},
+         year = 2018,
+        month = jun,
+       volume = {860},
+       number = {1},
+          eid = {41},
+        pages = {41},
+          doi = {10.3847/1538-4357/aac514},
+archivePrefix = {arXiv},
+       eprint = {1805.03206},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018ApJ...860...41L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@software{hstaXe,
+       author = {{Sosey}, Megan and {Bjkuhn} and {O'Steen}, Ricky and {Bradley}, Larry and {Tuong Nguyen}, Duy and {Som}, Debopam and {Hunkeler}, Joseph and {St-Apidgeon} and {Nphathi} and {Neuheisel}, Joshua and {Lim}, P.~L.},
+        title = "{spacetelescope/hstaxe: v1.0.5}",
+         year = 2023,
+        month = jul,
+          eid = {10.5281/zenodo.8136948},
+          doi = {10.5281/zenodo.8136948},
+      version = {v1.0.5},
+    publisher = {Zenodo},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023zndo...8136948S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 @MISC{pirzkal2020,
        author = {{Pirzkal}, Nor},
