+++ joss/paper.md
@@ -47,7 +47,7 @@ The Hubble Space Telescope Wide Field Camera 3 UV Imaging Spectrograph (HST WFC3
 
 # Statement of Need
 
-While the Hubble Space Telescope (HST) has been operating for more than 30 years, the WFC3-UVIS/G280 mode saw limited use until the recent surge of application to transmission spectroscopy of exoplanets [@wakeford2020; @lewis2020]. Current HST WFC3 pipelines [e.g. @eureka; @pacman] only service WFC-IR spectroscopy. HST WFC3-UVIS/G280 observations offer unique challenges in extracting spectral information: a curved spectral trace with a varied width, overlapping spectral orders, and high cosmic ray counts. These are in addition to potential variations in spectral extraction based on the use of different sub-array sizes and positions on the detector for each observation. Such challenges make current HST pipelines not suitable for reducing UVIS/G280 observations, and therefore a specialized pipeline is needed. 
+While the Hubble Space Telescope (HST) has been operating for more than 30 years, the WFC3-UVIS/G280 mode saw limited use until the recent surge of application to transmission spectroscopy of exoplanets [@wakeford2020; @lewis2020]. Most current HST WFC3 pipelines [e.g. @eureka; @pacman; @grizli] only service WFC-IR spectroscopy. Existing software tools for WFC3-UVIS data reduction are many years out of date [e.g. @hstaXe], restricted by licensed softwares such as IDL [e.g. @idl2; @idl2], or not specialized for use with this detector [e.g. @exotedrf]. HST WFC3-UVIS/G280 observations offer unique challenges in extracting spectral information: a curved spectral trace with a varied width, overlapping spectral orders, and high cosmic ray counts. These are in addition to potential variations in spectral extraction based on the use of different sub-array sizes and positions on the detector for each observation. Such challenges make current HST pipelines not suitable for reducing UVIS/G280 observations, and therefore a specialized, up-to-date, and fully open source pipeline is needed. 
 
 
 # Design and Features
