### 2026-04-24 6ffced6
visual diagnostics for stage_0.rst
