### 2026-04-24 cf84333
templated visuals for stage 2, to be added later
