+++ docs/source/views/tutorials/stage_2.rst
@@ -131,10 +131,130 @@ Then execute this script to run Stage 2! The output in your cell should look sim
 
 Assessing Stage 2's success
 ---------------------------
-Stage 2 is also a fairly customizable stage and has many diagnostics to look over. You will know if Stage 2 succeeded if:
+Stage 2 is also a fairly customizable stage and has many diagnostics to look over.
 
-  1. The calibration and aperture plots accurately located the correct traces and the apertures encompass the entirety of the orders you have chosen to extract from.
-  2. The extracted 1D spectral time series are consistent with the source spectrum expected (e.g. if your target is an A star, you should be able to see Balmer absorption lines in the spectrum centered at the appropriate wavelengths), and the 1D spectrum gif shows no cosmic ray spikes, drift over time, or reduction process artifacts (e.g. rapid variations or dramatic changes in flux in certain channels which can arise from hot/cold/dead pixels, an up-and-down jitteriness that might be the result of bad background subtraction, etc.).
-  3. The extracted orders are reasonably consistent with each other, where their wavelength ranges overlap. Some minor differences should be expected due to order throughput variations, but the overall shape and features should be comparable.
-  4. The cross-correlation dispersion and cross-dispersion shifts are reasonably consistent with the x-y shifts measured in the 0th order and background stars (if available) in Stage 1.
-  5. The extracted raw white light curves for each order clearly show the transit with the depth and scatter that you expect.
+
+Diagnostic 1: Calibration and apertures acquired traces
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+The calibration and aperture plots should show that Stage 2 located the correct traces, and that the apertures encompass the entirety of the orders you have chosen to extract from. We show aperture plots including the trace position calibration for the +1 and -1 traces below.
+
+.. list-table::
+   :widths: 256 256
+   :header-rows: 0
+
+   * - .. figure:: ../aperture+1.png
+          :width: 100%
+
+          *Aperture (top and bottom lines) encasing the +1 trace with a halfwidth of 12 pixels. The calibration solution (central line) traces the dispersed spectra to a sub-pixel accuracy.*
+     - .. figure:: ../aperture-1.png
+          :width: 100%
+
+          *Aperture (top and bottom lines) encasing the -1 trace with a halfwidth of 12 pixels. The calibration solution (central line) traces the dispersed spectra to a sub-pixel accuracy.*
+
+We additionally show calibration plots for orders +1 through +4, which can all be extracted albeit with increasing self-contamination.
+
+.. list-table::
+   :widths: 256 256 256 256
+   :header-rows: 0
+
+   * - .. figure:: ../calibration+1.png
+          :width: 100%
+
+          *The calibration solution for the +1 order.*
+     - .. figure:: ../calibration+2.png
+          :width: 100%
+
+          *The calibration solution for the +2 order.*
+     - .. figure:: ../calibration+3.png
+          :width: 100%
+
+          *The calibration solution for the +3 order.*
+     - .. figure:: ../calibration+4.png
+          :width: 100%
+
+          *The calibration solution for the +4 order.*
+
+Finally, we include calibration plots for orders -1 through -4, which can all be extracted albeit with increasing self-contamination.
+
+.. list-table::
+   :widths: 256 256 256 256
+   :header-rows: 0
+
+   * - .. figure:: ../calibration-1.png
+          :width: 100%
+
+          *The calibration solution for the -1 order.*
[…]
