### 2026-04-30 2b79a89
Update stage_2.rst
