+++ docs/source/views/tutorials/stage_2.rst
@@ -73,19 +73,20 @@ Step 2: Trace configuration
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 :code:`HUSTLE-tools` uses :code:`grismconf`, a package developed by `Pirzkal & Ryan 2017 <https://www.stsci.edu/files/live/sites/www/files/home/hst/instrumentation/wfc3/documentation/instrument-science-reports-isrs/_documents/2017/WFC3-2017-01.pdf>`_, to assign the field-dependent wavelength solution to the traces you want to extract. If you followed the instructions on the :doc:`Installation <views/installation>` page, you should have downloaded a set of `WFC3/UVIS configuration files <https://github.com/npirzkal/GRISM_WFC3>`_ which :code:`grismconf` will use to determine the wavelength solution for each order based on the target position you located in Stage 0 and optionally refined in Stage 1. Depending on which of the two WFC3/UVIS CCD chips you collected your observations on, you must supply the :code:`path_to_cal` variable with the absolute path to either your local UVIS_G280_CCD1_V2.conf file for chip 1, or your local UVIS_G280_CCD2_V2.conf file for chip 2. It is recommended by `Wakeford et al. 2020 <https://iopscience.iop.org/article/10.3847/1538-3881/ab7b78/pdf>`_ to use chip 2 for WFC3/UVIS observations because it is more stable than chip 1. Your observations are also most likely collected on chip 2, which you can confirm by opening the .fits files and checking the CCDCHIP keyword of the [SCI] header, as shown below.
 
-.. image:: ../ccdchipheader.png
-  :width: 256
+.. figure:: ../ccdchipheader.png
+  :width: 50%
   :align: left
-  :alt: [SCI] header info showing the CCD chip number stored in the CCDCHIP keyword
+  
+  *[SCI] header info showing the CCD chip number stored in the CCDCHIP keyword.*
 
 
-:code:`grismconf` supports extraction of both the positive (higher throughput, dispersed to left) orders and the negative (lower throughput, dispersed to right) orders, with support for up to the 4th order trace. You can specify which traces you want to extract by supplying the sign and order of each trace in the :code:`traces_to_conf` variable, e.g. by supplying this variable as ("+1","-1") you can request to extract the positive and negative 1st-order traces. Below we show a G280 data frame with the 0th order and all extractable traces marked.
+:code:`grismconf` supports extraction of both the positive (higher throughput, dispersed to left) orders and the negative (lower throughput, dispersed to right) orders, with support for up to the 4th order trace. All traces are extractable, but note that higher order traces have less flux and are more contaminated, which can make reliable extraction more challenging. You can specify which traces you want to extract by supplying the sign and order of each trace in the :code:`traces_to_conf` variable, e.g. by supplying this variable as ("+1","-1") you can request to extract the positive and negative 1st-order traces.
 
-.. image:: ../orders.png
-  :width: 512
+.. figure:: ../orders.png
+  :width: 100%
   :align: left
-  :alt: A WFC3/UVIS G280 data frame showing the 0th order (middle), all extractable positive orders (left), and all extractable negative orders (right)
-
+  
+  *A WFC3/UVIS G280 data frame showing the 0th order (middle), all extractable positive orders (left), and all extractable negative orders (right).*
 
 Shifts in the trace position over time can affect the wavelength solution, which is determined from the static direct photometric image taken at the observation start. You can refine the trace solution by setting :code:`refine_fit` to True to allow the trace solution to be updated in each frame.
 
@@ -145,11 +146,11 @@ The calibration and aperture plots should show that Stage 2 located the correct 
[…]
