### 2026-04-30 578708d
more plot features and some tutorial notes
