+++ docs/source/views/quick_start.rst
@@ -251,7 +251,13 @@ The outputs of the pipeline will be stored in :code:`output/outputs/`. Output pl
 4.1. The outputs of Stage 0
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-Stage 0 downloads our data and organizes it. The most important output of Stage 0 is the quicklookup.gif file. This gif compiles all of the G280 exposures, the total image flux, and the flux contained in an aperture containing the positive first order. We can use this gif to confirm that we captured the transit/eclipse of our target planet, or to look for issues such as tracking failures or satellite crossings. As you will see, the quicklookup.gif for this visit has no issues!
+Stage 0 downloads our data and organizes it. The most important output of Stage 0 is the quicklookup.gif file. This gif compiles all of the G280 exposures, the total image flux, and the flux contained in an aperture containing the positive first order. We can use this gif to confirm that we captured the transit/eclipse of our target planet, or to look for issues such as tracking failures or satellite crossings. As you can see below and in your own output folders, the quicklookup.gif for this visit has no issues!
+
+.. figure:: quicklookup.png
+  :width: 100%
+  :align: center
+  
+  *Last frame of quicklookup.gif file from visit 12 of HST-GO 17183, with a clear transit visible in the bottom right plot.*
 
 4.2. The outputs of Stage 1
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -265,6 +271,20 @@ Stage 1 handles data reduction, and outputs a lot of diagnostic plots we can use
 
 Stage 1 also outputs a revised quicklookup.gif that we can compare to our Stage 0 gif to see how the frames have changed after reduction. We also receive a similar gif showing the data quality array, which shows the pixels in each frame that were flagged for data quality issues during reduction (e.g. cosmic rays caught by temporal rejection, hot pixels caught by spatial rejection, etc.).
 
+.. list-table::
+   :widths: 512 512
+   :header-rows: 0
+
+   * - .. figure:: clean_quicklookup.png
+          :width: 100%
+
+          *The last frame of the quick lookup after cleaning. The data frames are free of cosmic rays and hot pixels with a smooth background (top), while the transit light curve shows greatly reduced scatter (bottom right).*
+     - .. figure:: dqquicklookup.png
+          :width: 100%
+
+          *The data quality array for the same frame. White marks pixels flagged as outliers. The total flags in the frame (bottom left) and the flags in the boxed region (right) are also displayed; neither exceeds 5\% of the data.*
+
+
 4.3. The outputs of Stage 2
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -276,7 +296,78 @@ Stage 2 extracts the 1D spectral time series from our reduced data frame, and li
   4. 1Dspec_order+1.png and 1Dspec_order-1.png show the first frame's 1D spectrum for each order. 2Dspec_order+1.png and 2Dspec_order-1.png plot the 1D spectra in every frame over time as a 2D map. 1Dspec_order+1.gif and 1Dspec_order-1.gif plays all extracted 1D spectra for each order as a gif. All of these plots can be used to assess the quality of the extracted spectra, including looking for uncorrected cosmic rays or systematic signals. In this dataset, a strong systematic can be seen in the -1 order at 400 nm.
[…]
