+++ joss/paper.bib
@@ -11,6 +11,36 @@
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{vandokkum2001,
+       author = {{van Dokkum}, Pieter G.},
+        title = "{Cosmic-Ray Rejection by Laplacian Edge Detection}",
+      journal = {\pasp},
+     keywords = {Instrumentation: Detectors, Methods: Data Analysis-techniques: image processing, Astrophysics},
+         year = 2001,
+        month = nov,
+       volume = {113},
+       number = {789},
+        pages = {1420-1427},
+          doi = {10.1086/323894},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0108003},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2001PASP..113.1420V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@MISC{pagul2023,
+       author = {{Pagul}, A. and {Ryan}, R. and {Kuhn}, B. and {Som}, D.},
+        title = "{The WFC3/UVIS G280 Grism Sky}",
+     keywords = {Hubble Space Telescope, HST, Space Telescope Science Institute, STScI, Wide Field Camera 3, WFC3},
+ howpublished = {Instrument Science Report WFC3 2023-06, 11 pages},
+         year = 2023,
+        month = sep,
+        pages = {6},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023wfc..rept....6P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{idl1,
        author = {{O'Meara}, John M. and {Prochaska}, J. Xavier and {Chen}, Hsiao-Wen and {Madau}, Piero},
         title = "{The Advanced Camera for Surveys+Wide Field Camera 3 Survey for Lyman Limit Systems. I. The Data}",
