+++ joss/paper.md
@@ -26,7 +26,7 @@ authors:
     affiliation: '1'
     corresponding: false
   - name: Ailsa M Campbell
-    orcid 0009-0006-4198-719X
+    orcid: 0009-0006-4198-719X
     affiliation: '3'
     corresponding: false
 affiliations:
