### 2026-04-30 9a60034
now with brighter higher orders
