### 2026-05-04 #9723 @AbbyBoehm
Hello all! We've recently pulled a patch to address your reviews! To break it down by review, first @mberg3:
1. We added GRIZLI to the citation list. We also address HSTaXe and the IDL code you mentioned, with comments in the paper as to why these are not the best options for the community (i.e. HSTaXe is out of date and IDL is not open source nor free).
2. The quickstart tutorial now has a simple script in it which will load the output spectra, sum it into a white light curve, bin and sum it into three spectroscopic light curves, and make a plot of all of these. It should hopefully serve as a good example to follow for those looking to bin their own light curves.
3. We've added a few remarks throughout the tutorials and JOSS paper re; all of these points, and we summarize those remarks here:
- For detecting the object and its traces, the only limiting factor is the quality of the photometric (usually F300X) image supplied, which is used to determine the wavelength solution in the G280 images. As long as the target is (1) bright enough to be clearly visible in an F300X image, and (2) dim enough to not saturate the F300X, GRISMCONF will have no trouble finding the traces.
- Higher orders imprint UV-blue flux onto the redder ends of the lower orders, so the higher orders really only become a factor if your star is very bright in the UV-blue compared to the red-IR, which will tend to happen if it is spectral type A or earlier. We have some example plots in the Stage 2 tutorial for traces 1-4 taken from an A star observation - the higher orders are very readily visible in those observations!
- Finally, GRISMCONF and HUSTLE-tools together can extract up to the +/-4th order, but there's not a lot of built-in support yet for dealing with spectral overlap. So these orders can be extracted in theory, but in practice you would only get useful flux from the wavelengths where they are not overlapped with the +1 order. Every order extracted is output separately into its own .nc file.
4. Tracing and extracting the -1 order pose no additional problems compared to the +1, since they rely only on the quality of the F300X image. Reduction has not posed additional problems for us so far, but in future updates we might consider detaching the + and - sigma thresholds from each other and allowing them to be tuned separately, as was done in Chachan et al. 2025.
5. We added a few extra details in the JOSS paper draft re; available treatment methods, but most of the detail on these methods will be found in the docs as well as in HUSTLE papers.

And for @radicamc:
1. exoTEDRF is now cited in our intro, with the remark that it is not specialized for the G280.
2. See (2) above. We did also link the docs to xarray in case users want to do some extra reading.
3. The .rst versions of the quickstart and tutorials on readthedocs now have plots. The .ipynb versions do not have these plots but they do have a link to the readthedocs for those looking to compare their figures. The plots are a little bit big at present, which has been noted in an open issue and which we will look into adjusting after a few other to-dos are resolved.
4. readthedocs now directly linked in the readme.
5. I don't think I have that power but I do appreciate the feedback :) I might repurpose that name or a similar variation of it someday, for another data reduction pipeline from an observatory whose name starts with H.

As always, please let me know if you have any comments and feedback regarding the latest build of the pipeline. We'll be happy to develop a patch to address any concerns! :)
