### 2026-06-09 da6e9c4
updated arxiv citation to aj citation for gascon 2025
