+++ joss/paper.bib
@@ -306,11 +306,20 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Gascon2025,
-      title={The HUSTLE Program: The UV to Near-IR Transmission Spectrum of the Hot Jupiter KELT-7b}, 
-      author={Carlos Gascón and Mercedes López-Morales and Ryan J. MacDonald and Joanna K. Barstow and Victoria A. Boehm and Hannah R. Wakeford and Munazza K. Alam and Lili Alderson and Natasha E. Batalha and Charlotte E. Fairman and David Grant and Nikole K. Lewis and Mark S. Marley and Sarah E. Moran and Kazumasa Ohno and Guillem Anglada-Escudé and Ignasi Ribas},
-      year={2025},
-      eprint={2506.19193},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.EP},
-      url={https://arxiv.org/abs/2506.19193}, 
+       author = {{Gasc{\'o}n}, C. and {L{\'o}pez-Morales}, M. and {MacDonald}, R.~J. and {Barstow}, J.~K. and {Boehm}, V.~A. and {Wakeford}, H.~R. and {Alam}, M.~K. and {Alderson}, L. and {Batalha}, N.~E. and {Fairman}, C.~E. and {Grant}, D. and {Lewis}, N.~K. and {Marley}, M.~S. and {Moran}, S.~E. and {Ohno}, K. and {Anglada-Escud{\'e}}, G. and {Ribas}, I.},
+        title = "{The HUSTLE Program: The UV to Near-IR Transmission Spectrum of the Hot Jupiter KELT-7b}",
+      journal = {\aj},
+     keywords = {Exoplanets, Hot Jupiters, Exoplanet atmospheres, 498, 753, 487, Earth and Planetary Astrophysics},
+         year = 2025,
+        month = sep,
+       volume = {170},
+       number = {3},
+          eid = {178},
+        pages = {178},
+          doi = {10.3847/1538-3881/ade67e},
+archivePrefix = {arXiv},
+       eprint = {2506.19193},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025AJ....170..178G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
