### 2026-06-23 d43a77c
removed journal shortcuts, added software citations
