+++ joss/paper.bib
@@ -14,7 +14,7 @@
 @ARTICLE{vandokkum2001,
        author = {{van Dokkum}, Pieter G.},
         title = "{Cosmic-Ray Rejection by Laplacian Edge Detection}",
-      journal = {\pasp},
+      journal = {Publications of the Astronomical Society of the Pacific},
      keywords = {Instrumentation: Detectors, Methods: Data Analysis-techniques: image processing, Astrophysics},
          year = 2001,
         month = nov,
@@ -44,7 +44,7 @@ archivePrefix = {arXiv},
 @ARTICLE{idl1,
        author = {{O'Meara}, John M. and {Prochaska}, J. Xavier and {Chen}, Hsiao-Wen and {Madau}, Piero},
         title = "{The Advanced Camera for Surveys+Wide Field Camera 3 Survey for Lyman Limit Systems. I. The Data}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplement},
      keywords = {intergalactic medium, methods: data analysis, quasars: absorption lines, quasars: general, surveys, ultraviolet: general},
          year = 2011,
         month = aug,
@@ -60,7 +60,7 @@ archivePrefix = {arXiv},
 @ARTICLE{idl2,
        author = {{Lusso}, Elisabeta and {Fumagalli}, Michele and {Rafelski}, Marc and {Neeleman}, Marcel and {Prochaska}, Jason X. and {Hennawi}, Joseph F. and {O'Meara}, John M. and {Theuns}, Tom},
         title = "{The Spectral and Environment Properties of z {\ensuremath{\sim}} 2.0-2.5 Quasar Pairs}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {accretion, accretion disks, galaxies: active, intergalactic medium, quasars: absorption lines, quasars: general, Astrophysics - Astrophysics of Galaxies},
          year = 2018,
         month = jun,
@@ -104,7 +104,7 @@ archivePrefix = {arXiv},
 @ARTICLE{ohno2020,
        author = {{Ohno}, Kazumasa and {Kawashima}, Yui},
         title = "{Super-Rayleigh Slopes in Transmission Spectra of Exoplanets Generated by Photochemical Haze}",
-      journal = {\apjl},
+      journal = {The Astrophysical Journal Letters},
      keywords = {Exoplanet atmospheres, Hot Jupiters, Optical observation, Transmission spectroscopy, 487, 753, 1169, 2133, Astrophysics - Earth and Planetary Astrophysics},
          year = 2020,
         month = jun,
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
 @ARTICLE{bourrier2018,
        author = {{Bourrier}, V. and {Lecavelier des Etangs}, A. and {Ehrenreich}, D. and {Sanz-Forcada}, J. and {Allart}, R. and {Ballester}, G.~E. and {Buchhave}, L.~A. and {Cohen}, O. and {Deming}, D. and {Evans}, T.~M. and {Garc{\'\i}a Mu{\~n}oz}, A. and {Henry}, G.~W. and {Kataria}, T. and {Lavvas}, P. and {Lewis}, N. and {L{\'o}pez-Morales}, M. and {Marley}, M. and {Sing}, D.~K. and {Wakeford}, H.~R.},
         title = "{Hubble PanCET: an extended upper atmosphere of neutral hydrogen around the warm Neptune GJ 3470b}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
[…]
