### 2026-06-23 6e5805c
sentence reworks and extended acknowledgments
