+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'HUSTLE-tools: a one stop shop for Hubble WFC3-UVIS/G280 spectral reduction'
+title: 'HUSTLE-tools: A One Stop Shop For Hubble WFC3-UVIS/G280 Spectral Reduction'
 tags:
   - Python
   - astronomy
@@ -44,14 +44,14 @@ bibliography: paper.bib
 
 # Summary
 
-Fully understanding the complex physical and chemical processes that shape exoplanet atmospheres requires the complete spectrum from the ultraviolet (UV) to the mid-infrared (IR). In particular, the UV-optical has proven vital to constraining the presence of UV absorbers or scatterers [e.g. @lothringer2022], atmospheric escape [e.g. @bourrier2018], and enhanced scattering due to aerosol opacities [@ohno2020].
+Fully understanding the complex physical and chemical processes that shape exoplanet atmospheres requires the complete spectrum from the ultraviolet (UV) to the mid-infrared (IR). In particular, UV-optical observations have proven vital for identifying UV-absorbing atmospheric species [e.g. @lothringer2022], detecting atmospheric escape [e.g. @bourrier2018], and constraining the presence of aerosols through their opacity and enhanced scattering of UV-optical light [@ohno2020].
 
-The Hubble Space Telescope Wide Field Camera 3 UV Imaging Spectrograph (HST WFC3-UVIS) G280 grism is the only low-resolution spectrograph that gives us access to UV-optical wavelengths (0.2--0.8 microns) simultaneously. With most current instrumentation operating in near- and mid-IR wavelengths, and UV successors to HST still many years away, HST WFC3-UVIS/G280 emerges as the only instrument capable of unveiling the physics and chemistry that takes place in the upper atmospheric layers of exoplanets. Developed in the context of the Hubble Ultraviolet-optical Survey of Transiting Legacy Exoplanets (HUSTLE, GO-17183 PI: Wakeford), HUSTLE-tools is an open-source package of modules designed to easily download, organize, analyze, clean, and extract the target spectrum from HST WFC3-UVIS/G280 spectral images.
+The Hubble Space Telescope Wide Field Camera 3 UV Imaging Spectrograph (HST WFC3-UVIS) G280 grism is the only low-resolution spectrograph that gives us access to UV-optical wavelengths (0.2--0.8 microns) simultaneously. With most current instrumentation operating in near- and mid-IR wavelengths, and UV successors to HST still many years away, HST WFC3-UVIS/G280 emerges as the only instrument capable of unveiling the physics and chemistry that takes place in the upper atmospheric layers of exoplanets. Developed in the context of the Hubble Ultraviolet-optical Survey of Transiting Legacy Exoplanets (HUSTLE, GO-17183 PI: Wakeford), HUSTLE-tools is an open-source package of modules designed to easily download, organize, analyze, clean, and extract the target spectra from HST WFC3-UVIS/G280 spectral images.
 
 
 # Statement of Need
 
[…]
